refactor(db): repository_pg 拆分为多子模块包
将 2928 行的 repository_pg.py 拆为 schema/lifecycle/context/message/blacklist/admin/config/image_cache 子模块，经 __init__ 全量重导出保持导入面。子模块经包命名空间在调用时解析共享运行时符号，保证测试字符串打桩传播。SQL/ORM/索引/迁移语义不变。
