--- pallas/core/foundation/db/repository_pg.py
@@ -1,2931 +1,0 @@
-"""PostgreSQL Repository 实现"""
-
-from __future__ import annotations
-
-import asyncio
-import contextlib
-import hashlib
-import json
-import os
-import time
-from collections import OrderedDict
-from contextlib import asynccontextmanager
-from typing import TYPE_CHECKING, Any
-
-from nonebot import logger
-from sqlalchemy import (
-    JSON,
-    BigInteger,
-    Boolean,
-    Float,
-    ForeignKey,
-    Index,
-    Integer,
-    LargeBinary,
-    Text,
-    UniqueConstraint,
-    and_,
-    case,
-    delete,
-    func,
-    insert,
-    inspect,
-    literal_column,
-    or_,
-    select,
-    text,
-    update,
-)
-from sqlalchemy.dialects.postgresql import JSONB
-from sqlalchemy.dialects.postgresql import insert as pg_insert
-from sqlalchemy.exc import IntegrityError
-from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession, async_sessionmaker
-from sqlalchemy.orm import DeclarativeBase, Mapped, aliased, mapped_column, relationship, selectinload
-
-from pallas.core.foundation.db.blob_store import (
-    delete_image_blob,
-    image_blob_rel_path,
-    read_image_blob_at,
-    write_image_blob,
-)
-from pallas.core.platform.observability import slow_path_threshold_ms
-
-if TYPE_CHECKING:
-    from pallas.core.foundation.db.modules import Answer, Ban, BlackList, Context, ImageCache, Message
-    from pallas.core.foundation.db.repository import ImageCachePrunePolicy, ImageCachePruneResult
-
-_JsonB = JSONB().with_variant(JSON(), "sqlite")
-
-
-# ---------------------------------------------------------------------------
-# 运行时 \x00 防御：PostgreSQL TEXT 不接受 NUL 字符，需在写入前统一剥除
-# ---------------------------------------------------------------------------
-
-
-def _s(x: str | None) -> str | None:
-    if x is None:
-        return None
-    return x.replace("\x00", "") if "\x00" in x else x
-
-
-def _strip_null_deep(obj: Any) -> Any:
-    """递归剥除 str / dict / list 中的 \\u0000，用于 JSONB 字段。"""
-    if isinstance(obj, str):
-        return _s(obj)
-    if isinstance(obj, dict):
-        return {k: _strip_null_deep(v) for k, v in obj.items()}
-    if isinstance(obj, list):
-        return [_strip_null_deep(i) for i in obj]
-    return obj
-
-
-# ---------------------------------------------------------------------------
-# ORM 定义
-# ---------------------------------------------------------------------------
-
-
-class Base(DeclarativeBase):
-    pass
-
-
-class ContextAnswerRow(Base):
-    __tablename__ = "context_answer"
-    # keywords_hash 定长 md5 唯一索引
-    __table_args__ = (
-        UniqueConstraint("context_id", "group_id", "keywords_hash", name="uq_context_answer_ctx_group_kw"),
-        Index("ix_context_answer_ctx_count_time", "context_id", "count", "time"),
-    )
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    context_id: Mapped[int] = mapped_column(ForeignKey("context.id", ondelete="CASCADE"), nullable=False)
-    keywords: Mapped[str] = mapped_column(Text, nullable=False)
-    keywords_hash: Mapped[str] = mapped_column(Text, nullable=False)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    count: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
-    time: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-
-    messages: Mapped[list[ContextAnswerMessageRow]] = relationship(
-        "ContextAnswerMessageRow", cascade="all, delete-orphan", lazy="noload"
-    )
-
-
-class ContextAnswerMessageRow(Base):
-    __tablename__ = "context_answer_message"
-    __table_args__ = (Index("ix_context_answer_message_answer_id_id", "answer_id", "id"),)
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    answer_id: Mapped[int] = mapped_column(
-        ForeignKey("context_answer.id", ondelete="CASCADE"), nullable=False, index=True
-    )
-    message: Mapped[str] = mapped_column(Text, nullable=False)
-
-
-class ContextBanRow(Base):
-    __tablename__ = "context_ban"
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    context_id: Mapped[int] = mapped_column(ForeignKey("context.id", ondelete="CASCADE"), nullable=False, index=True)
-    keywords: Mapped[str] = mapped_column(Text, nullable=False)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    reason: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    time: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-
-
-class ContextRow(Base):
-    __tablename__ = "context"
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    keywords: Mapped[str] = mapped_column(Text, nullable=False)
-    # unique=True 已由 PG 自动建 btree，不再附加 index=True 以免冗余索引
-    keywords_hash: Mapped[str] = mapped_column(Text, nullable=False, unique=True)
-    time: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    trigger_count: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
-    clear_time: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-
-    answers: Mapped[list[ContextAnswerRow]] = relationship(
-        "ContextAnswerRow", cascade="all, delete-orphan", lazy="noload"
-    )
-    ban: Mapped[list[ContextBanRow]] = relationship("ContextBanRow", cascade="all, delete-orphan", lazy="noload")
-
-
-class MessageRow(Base):
-    __tablename__ = "message"
-    __table_args__ = (
-        Index("ix_message_time", "time"),
-        Index("ix_message_group_time", "group_id", "time"),
-        Index("ix_message_group_user_time", "group_id", "user_id", "time"),
-        # 消息幂等落库锚点：同 (group_id, bot_id, message_id) 只留一条。
-        # message_id 为 NULL 的行不参与冲突（PG 中 NULL <> NULL，互不冲突）。
-        Index("uq_message_group_bot_message_id", "group_id", "bot_id", "message_id", unique=True),
-    )
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
-    user_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    bot_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    raw_message: Mapped[str] = mapped_column(Text, nullable=False)
-    is_plain_text: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    plain_text: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    keywords: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    sender_name: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    message_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
-    reply_to_message_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
-    suppressed_by_rage: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    time: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-
-
-class BackgroundJobRow(Base):
-    __tablename__ = "background_job"
-    __table_args__ = (
-        UniqueConstraint("idempotency_key", name="uq_background_job_idempotency"),
-        Index("ix_background_job_claim", "status", "available_at", "leased_until", "id"),
-        Index("ix_background_job_delivery_claim", "kind", "status", "finished_at"),
-    )
-
-    id: Mapped[str] = mapped_column(Text, primary_key=True)
-    kind: Mapped[str] = mapped_column(Text, nullable=False)
-    payload: Mapped[Any] = mapped_column(_JsonB, nullable=False)
-    idempotency_key: Mapped[str] = mapped_column(Text, nullable=False)
-    status: Mapped[str] = mapped_column(Text, nullable=False, default="pending")
-    attempts: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    available_at: Mapped[float] = mapped_column(Float, nullable=False, default=0)
-    leased_until: Mapped[float | None] = mapped_column(Float, nullable=True)
-    lease_owner: Mapped[str | None] = mapped_column(Text, nullable=True)
-    lease_id: Mapped[str | None] = mapped_column(Text, nullable=True)
-    last_error: Mapped[str | None] = mapped_column(Text, nullable=True)
-    created_at: Mapped[float] = mapped_column(Float, nullable=False, default=0)
-    finished_at: Mapped[float | None] = mapped_column(Float, nullable=True)
-
-
-class BlackListRow(Base):
-    __tablename__ = "blacklist"
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False, unique=True)
-    answers: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=list)
-    answers_reserve: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=list)
-
-
-class AdminMemberRow(Base):
-    __tablename__ = "admin_members"
-    __table_args__ = (
-        UniqueConstraint("scope", "bot_id", "user_id", name="uq_admin_members_scope_bot_user"),
-        Index("ix_admin_members_bot_user", "bot_id", "user_id"),
-    )
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    scope: Mapped[str] = mapped_column(Text, nullable=False)
-    bot_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
-    user_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    note: Mapped[str | None] = mapped_column(Text, nullable=True)
-    created_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    updated_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-
-
-class PallasACLRow(Base):
-    __tablename__ = "acl_rules"
-    __table_args__ = (
-        UniqueConstraint(
-            "role",
-            "subject",
-            "action",
-            "target_scope",
-            "target",
-            name="uq_acl_rules_signature",
-        ),
-        Index("ix_acl_rules_action", "action"),
-    )
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    role: Mapped[str] = mapped_column(Text, nullable=False)
-    subject: Mapped[str | None] = mapped_column(Text, nullable=True)
-    action: Mapped[str] = mapped_column(Text, nullable=False)
-    target_scope: Mapped[str] = mapped_column(Text, nullable=False)
-    target: Mapped[str] = mapped_column(Text, nullable=False)
-    effect: Mapped[str] = mapped_column(Text, nullable=False)
-    priority: Mapped[int] = mapped_column(Integer, nullable=False, default=100)
-    source: Mapped[str] = mapped_column(Text, nullable=False, default="user")
-    created_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    updated_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-
-
-class SchemaMigrationRow(Base):
-    __tablename__ = "schema_migrations"
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    step: Mapped[str] = mapped_column(Text, nullable=False, unique=True)
-    applied_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-
-
-class BotConfigRow(Base):
-    __tablename__ = "bot_config"
-
-    account: Mapped[int] = mapped_column(BigInteger, primary_key=True)
-    admins: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=list)
-    auto_accept_friend: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    auto_accept_group: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    security: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    taken_name: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=dict)
-    drunk: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=dict)
-    disabled_plugins: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=list)
-    disabled_plugins_audit: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=list)
-    community_roster_show_qq: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
-    persona: Mapped[Any] = mapped_column(_JsonB, nullable=True)
-    group_style_enabled: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
-    plugin_storage: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=dict)
-
-
-class GroupConfigRow(Base):
-    __tablename__ = "group_config"
-
-    group_id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
-    roulette_mode: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
-    banned: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    sing_progress: Mapped[Any] = mapped_column(_JsonB, nullable=True)
-    disabled_plugins: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=list)
-    disabled_plugins_audit: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=list)
-    blocked_user_ids: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=list)
-    style_profile: Mapped[Any] = mapped_column(_JsonB, nullable=True)
-    plugin_storage: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=dict)
-
-
-class UserConfigRow(Base):
-    __tablename__ = "user_config"
-
-    user_id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
-    banned: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    banned_by: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    banned_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    maa_devices: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=dict)
-    maa_active_device: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    maa_stage_plan: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=list)
-    plugin_storage: Mapped[Any] = mapped_column(_JsonB, nullable=False, default=dict)
-
-
-class ImageCacheRow(Base):
-    __tablename__ = "image_cache"
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    cq_code: Mapped[str] = mapped_column(Text, unique=True, nullable=False)
-    content_hash: Mapped[str | None] = mapped_column(Text, nullable=True, index=True)
-    # 原生二进制 BYTEA。旧库 base64_data 由 init_pg → _ensure_pg_image_cache_blob_data 幂等迁移。
-    blob_data: Mapped[bytes | None] = mapped_column(LargeBinary, nullable=True)
-    # 二进制已落文件（data/image_cache_blobs），DB 只存相对路径与字节数。
-    blob_path: Mapped[str | None] = mapped_column(Text, nullable=True)
-    blob_size: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
-    ref_times: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
-    date: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
-
-
-class StickerLabelRow(Base):
-    """按原始二进制内容哈希缓存的表情语义标签。"""
-
-    __tablename__ = "sticker_label"
-
-    content_hash: Mapped[str] = mapped_column(Text, primary_key=True)
-    is_sticker: Mapped[bool] = mapped_column(Boolean, nullable=False)
-    confidence: Mapped[float] = mapped_column(Float, nullable=False)
-    prompt_version: Mapped[int] = mapped_column(Integer, nullable=False)
-    labeled_at: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
-    label_json: Mapped[Any] = mapped_column(_JsonB, nullable=False)
-
-
-class UserStickerStatRow(Base):
-    """群成员发送图片的次数统计，按内容哈希聚合，不带 bot 维度。"""
-
-    __tablename__ = "user_sticker_stat"
-    __table_args__ = (
-        UniqueConstraint("group_id", "user_id", "content_hash", name="uq_user_sticker_stat_group_user_hash"),
-        Index("ix_user_sticker_stat_send_count", "send_count"),
-    )
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    user_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    content_hash: Mapped[str] = mapped_column(Text, nullable=False)
-    send_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    last_sent_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    updated_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-
-
-class LlmChatMessageRow(Base):
-    __tablename__ = "llm_chat_message"
-    __table_args__ = (
-        Index("ix_llm_chat_message_bot_group_time", "bot_id", "group_id", "created_at"),
-        Index("ix_llm_chat_message_bot_group_user_time", "bot_id", "group_id", "user_id", "created_at"),
-    )
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    bot_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    user_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    role: Mapped[str] = mapped_column(Text, nullable=False)
-    content: Mapped[str] = mapped_column(Text, nullable=False)
-    created_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-
-
-class LlmMemoryEntryRow(Base):
-    __tablename__ = "llm_memory_entry"
-    __table_args__ = (Index("ix_llm_memory_entry_bot_group_time", "bot_id", "group_id", "updated_at"),)
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    bot_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    keywords: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    content: Mapped[str] = mapped_column(Text, nullable=False)
-    source: Mapped[str] = mapped_column(Text, nullable=False, default="teach")
-    importance: Mapped[float] = mapped_column(Float, nullable=False, default=0.5)
-    confidence: Mapped[float] = mapped_column(Float, nullable=False, default=0.5)
-    expires_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    visibility: Mapped[str] = mapped_column(Text, nullable=False, default="private")
-    embedding_json: Mapped[str | None] = mapped_column(Text, nullable=True, default=None)
-    embedding_model: Mapped[str | None] = mapped_column(Text, nullable=True, default=None)
-    created_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    updated_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-
-
-class LlmRelationshipNoteRow(Base):
-    """关系备注层：按 (bot, group, user) 维护稳定关系事实，带置信权重与衰减。"""
-
-    __tablename__ = "llm_relationship_note"
-    __table_args__ = (
-        Index(
-            "ix_llm_relationship_note_scope",
-            "bot_id",
-            "group_id",
-            "user_id",
-            unique=True,
-        ),
-    )
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    bot_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    user_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    content: Mapped[str] = mapped_column(Text, nullable=False)
-    source: Mapped[str] = mapped_column(Text, nullable=False, default="teach")
-    weight: Mapped[float] = mapped_column(Float, nullable=False, default=1.0)
-    warmth_delta: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
-    assertiveness_delta: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
-    affinity: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
-    rage: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    rage_last_attack_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    rage_last_attack_message_id: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    rage_silenced_until: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    rage_silence_reason: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    created_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    updated_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-
-
-class LlmMemoryEntityRow(Base):
-    """记忆图谱实体（人/概念等）。"""
-
-    __tablename__ = "llm_memory_entity"
-    __table_args__ = (
-        Index("uq_llm_memory_entity_scope_name", "scope_key", "name", unique=True),
-        Index("ix_llm_memory_entity_bot_group_time", "bot_id", "group_id", "updated_at"),
-    )
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    scope_key: Mapped[str] = mapped_column(Text, nullable=False)
-    bot_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    name: Mapped[str] = mapped_column(Text, nullable=False)
-    summary: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    tags_json: Mapped[str] = mapped_column(Text, nullable=False, default="[]")
-    kind: Mapped[str] = mapped_column(Text, nullable=False, default="concept")
-    user_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True, default=None)
-    source: Mapped[str] = mapped_column(Text, nullable=False, default="manual")
-    deleted_at: Mapped[int | None] = mapped_column(BigInteger, nullable=True, default=None)
-    created_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    updated_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-
-
-class LlmMemoryEdgeRow(Base):
-    """记忆图谱关系边。"""
-
-    __tablename__ = "llm_memory_edge"
-    __table_args__ = (
-        Index("ix_llm_memory_edge_scope_valid", "scope_key", "valid_at"),
-        Index("ix_llm_memory_edge_scope_source", "scope_key", "source_entity_id"),
-        Index("ix_llm_memory_edge_scope_target", "scope_key", "target_entity_id"),
-        Index("ix_llm_memory_edge_bot_group", "bot_id", "group_id"),
-    )
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    scope_key: Mapped[str] = mapped_column(Text, nullable=False)
-    bot_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    fact: Mapped[str] = mapped_column(Text, nullable=False)
-    source_entity_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    target_entity_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    relation_type: Mapped[str] = mapped_column(Text, nullable=False, default="related_to")
-    weight: Mapped[float] = mapped_column(Float, nullable=False, default=1.0)
-    mention_count: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
-    episode_ids_json: Mapped[str] = mapped_column(Text, nullable=False, default="[]")
-    valid_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    invalid_at: Mapped[int | None] = mapped_column(BigInteger, nullable=True, default=None)
-    source: Mapped[str] = mapped_column(Text, nullable=False, default="manual")
-    created_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    updated_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-
-
-class LlmMemoryCategoryRow(Base):
-    """记忆图谱分类（分层语义树节点）。"""
-
-    __tablename__ = "llm_memory_category"
-    __table_args__ = (
-        Index("uq_llm_memory_category_scope_layer_name", "scope_key", "layer", "name", unique=True),
-        Index("ix_llm_memory_category_bot_group", "bot_id", "group_id"),
-    )
-
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    scope_key: Mapped[str] = mapped_column(Text, nullable=False)
-    bot_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    name: Mapped[str] = mapped_column(Text, nullable=False)
-    summary: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    tags_json: Mapped[str] = mapped_column(Text, nullable=False, default="[]")
-    layer: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
-    parent_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True, default=None)
-    member_entity_ids_json: Mapped[str] = mapped_column(Text, nullable=False, default="[]")
-    source: Mapped[str] = mapped_column(Text, nullable=False, default="manual")
-    deleted_at: Mapped[int | None] = mapped_column(BigInteger, nullable=True, default=None)
-    created_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    updated_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
-
-
-class LlmMemoryHierStatusRow(Base):
-    """分层语义图重建状态（每 scope 一行）。"""
-
-    __tablename__ = "llm_memory_hier_status"
-
-    scope_key: Mapped[str] = mapped_column(Text, primary_key=True)
-    bot_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    group_id: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    max_layer: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    last_rebuild_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-    entity_count_at_rebuild: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    group_summary: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    updated_at: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
-
-
-# ---------------------------------------------------------------------------
-# 引擎 / 会话
-# ---------------------------------------------------------------------------
-
-
-_engine: AsyncEngine | None = None
-_session_factory: async_sessionmaker[AsyncSession] | None = None
-_reply_query_snapshot_cache: OrderedDict[str, tuple[float, Context | None]] = OrderedDict()
-_reply_query_snapshot_inflight: dict[str, asyncio.Task[Context | None]] = {}
-_reply_query_snapshot_lock = asyncio.Lock()
-_DELETE_ID_BATCH = 1000
-
-
-def _ensure_pg_image_cache_blob_data(connection) -> None:
-    """旧库 image_cache 仍为 base64_data(TEXT) 时迁到 blob_data(BYTEA)。
-
-    幂等，可重复调用。``create_all`` 不会改已有表列，故升级路径必须走这里。
-    """
-    insp = inspect(connection)
-    if not insp.has_table("image_cache"):
-        return
-    names = {c["name"] for c in insp.get_columns("image_cache")}
-    if "blob_data" in names and "base64_data" not in names:
-        return
-    if "blob_data" not in names and "base64_data" not in names:
-        connection.execute(text("ALTER TABLE image_cache ADD COLUMN blob_data BYTEA"))
-        return
-    if "blob_data" not in names:
-        connection.execute(text("ALTER TABLE image_cache ADD COLUMN blob_data BYTEA"))
-        logger.info("image_cache: 已添加 blob_data 列，开始从 base64_data 迁移")
-    if "base64_data" in names:
-        connection.execute(
-            text(
-                "UPDATE image_cache SET blob_data = decode(base64_data, 'base64') "
-                "WHERE base64_data IS NOT NULL AND blob_data IS NULL"
-            )
-        )
-        connection.execute(text("ALTER TABLE image_cache DROP COLUMN base64_data"))
-        logger.info("image_cache: base64_data → blob_data 迁移完成")
-
-
-def _ensure_pg_image_cache_content_hash(connection) -> None:
-    insp = inspect(connection)
-    if not insp.has_table("image_cache"):
-        return
-    names = {c["name"] for c in insp.get_columns("image_cache")}
-    if "content_hash" not in names:
-        connection.execute(text("ALTER TABLE image_cache ADD COLUMN content_hash TEXT"))
-    connection.execute(text("CREATE INDEX IF NOT EXISTS ix_image_cache_content_hash ON image_cache (content_hash)"))
-
-
-def _ensure_pg_image_cache_blob_path(connection) -> None:
-    """image_cache 二进制落文件后补 blob_path/blob_size 元数据列。幂等。"""
-    insp = inspect(connection)
-    if not insp.has_table("image_cache"):
-        return
-    names = {c["name"] for c in insp.get_columns("image_cache")}
-    if "blob_path" not in names:
-        connection.execute(text("ALTER TABLE image_cache ADD COLUMN blob_path TEXT"))
-    if "blob_size" not in names:
-        connection.execute(text("ALTER TABLE image_cache ADD COLUMN blob_size BIGINT"))
-
-
-def _ensure_pg_group_config_blocked_user_ids(connection) -> None:
-    """旧库 group_config 缺列时补列。"""
-    insp = inspect(connection)
-    if not insp.has_table("group_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("group_config")}
-    if "blocked_user_ids" in names:
-        return
-    connection.execute(text("ALTER TABLE group_config ADD COLUMN blocked_user_ids JSONB NOT NULL DEFAULT '[]'::jsonb"))
-
-
-def _ensure_pg_background_job_lease_id(connection) -> None:
-    insp = inspect(connection)
-    if not insp.has_table("background_job"):
-        return
-    names = {column["name"] for column in insp.get_columns("background_job")}
-    if "lease_id" not in names:
-        connection.execute(text("ALTER TABLE background_job ADD COLUMN lease_id TEXT"))
-    if "last_error" not in names:
-        connection.execute(text("ALTER TABLE background_job ADD COLUMN last_error TEXT"))
-
-
-def _ensure_pg_group_config_style_profile(connection) -> None:
-    """旧库 group_config 缺列时补列。"""
-    insp = inspect(connection)
-    if not insp.has_table("group_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("group_config")}
-    if "style_profile" in names:
-        return
-    connection.execute(text("ALTER TABLE group_config ADD COLUMN style_profile JSONB"))
-
-
-def _ensure_pg_group_config_plugin_storage(connection) -> None:
-    insp = inspect(connection)
-    if not insp.has_table("group_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("group_config")}
-    if "plugin_storage" in names:
-        return
-    connection.execute(text("ALTER TABLE group_config ADD COLUMN plugin_storage JSONB NOT NULL DEFAULT '{}'::jsonb"))
-
-
-def _ensure_pg_group_config_disabled_plugins_audit(connection) -> None:
-    """旧库 group_config 缺审计列时补列。"""
-    insp = inspect(connection)
-    if not insp.has_table("group_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("group_config")}
-    if "disabled_plugins_audit" in names:
-        return
-    connection.execute(
-        text("ALTER TABLE group_config ADD COLUMN disabled_plugins_audit JSONB NOT NULL DEFAULT '[]'::jsonb")
-    )
-
-
-def _ensure_pg_bot_config_disabled_plugins_audit(connection) -> None:
-    """旧库 bot_config 缺审计列时补列。"""
-    insp = inspect(connection)
-    if not insp.has_table("bot_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("bot_config")}
-    if "disabled_plugins_audit" in names:
-        return
-    connection.execute(
-        text("ALTER TABLE bot_config ADD COLUMN disabled_plugins_audit JSONB NOT NULL DEFAULT '[]'::jsonb")
-    )
-
-
-def _ensure_pg_bot_config_persona(connection) -> None:
-    """旧库 bot_config 缺列时补列。"""
-    insp = inspect(connection)
-    if not insp.has_table("bot_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("bot_config")}
-    if "persona" in names:
-        return
-    connection.execute(text("ALTER TABLE bot_config ADD COLUMN persona JSONB"))
-
-
-def _ensure_pg_bot_config_community_roster_show_qq(connection) -> None:
-    """旧库 bot_config 缺列时补列。"""
-    insp = inspect(connection)
-    if not insp.has_table("bot_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("bot_config")}
-    if "community_roster_show_qq" in names:
-        return
-    connection.execute(text("ALTER TABLE bot_config ADD COLUMN community_roster_show_qq BOOLEAN NOT NULL DEFAULT true"))
-
-
-def _ensure_pg_bot_config_group_style_enabled(connection) -> None:
-    """旧库 bot_config 缺列时补列。"""
-    insp = inspect(connection)
-    if not insp.has_table("bot_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("bot_config")}
-    if "group_style_enabled" in names:
-        return
-    connection.execute(text("ALTER TABLE bot_config ADD COLUMN group_style_enabled BOOLEAN NOT NULL DEFAULT true"))
-
-
-def _ensure_pg_user_config_maa_devices(connection) -> None:
-    """旧库 user_config 缺列时补列。"""
-    insp = inspect(connection)
-    if not insp.has_table("user_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("user_config")}
-    if "maa_devices" not in names:
-        connection.execute(text("ALTER TABLE user_config ADD COLUMN maa_devices JSONB NOT NULL DEFAULT '{}'::jsonb"))
-    if "maa_active_device" not in names:
-        connection.execute(text("ALTER TABLE user_config ADD COLUMN maa_active_device TEXT NOT NULL DEFAULT ''"))
-    if "maa_stage_plan" not in names:
-        connection.execute(text("ALTER TABLE user_config ADD COLUMN maa_stage_plan JSONB NOT NULL DEFAULT '[]'::jsonb"))
-    if "plugin_storage" not in names:
-        connection.execute(text("ALTER TABLE user_config ADD COLUMN plugin_storage JSONB NOT NULL DEFAULT '{}'::jsonb"))
-
-
-def _ensure_pg_user_config_ban_audit(connection) -> None:
-    """旧库 user_config 缺拉黑审计列时补列。"""
-    insp = inspect(connection)
-    if not insp.has_table("user_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("user_config")}
-    if "banned_by" not in names:
-        connection.execute(text("ALTER TABLE user_config ADD COLUMN banned_by TEXT NOT NULL DEFAULT ''"))
-    if "banned_at" not in names:
-        connection.execute(text("ALTER TABLE user_config ADD COLUMN banned_at BIGINT NOT NULL DEFAULT 0"))
-
-
-def _ensure_pg_bot_config_plugin_storage(connection) -> None:
-    insp = inspect(connection)
-    if not insp.has_table("bot_config"):
-        return
-    names = {c["name"] for c in insp.get_columns("bot_config")}
-    if "plugin_storage" in names:
-        return
-    connection.execute(text("ALTER TABLE bot_config ADD COLUMN plugin_storage JSONB NOT NULL DEFAULT '{}'::jsonb"))
-
-
-def _ensure_pg_llm_relationship_delta_columns(connection) -> None:
-    """旧库 llm_relationship_note 缺人对语气偏置列时补列。"""
-    insp = inspect(connection)
-    if not insp.has_table("llm_relationship_note"):
-        return
-    names = {c["name"] for c in insp.get_columns("llm_relationship_note")}
-    if "warmth_delta" not in names:
-        connection.execute(
-            text("ALTER TABLE llm_relationship_note ADD COLUMN warmth_delta DOUBLE PRECISION NOT NULL DEFAULT 0")
-        )
-    if "assertiveness_delta" not in names:
-        connection.execute(
-            text("ALTER TABLE llm_relationship_note ADD COLUMN assertiveness_delta DOUBLE PRECISION NOT NULL DEFAULT 0")
-        )
-    if "affinity" not in names:
-        connection.execute(
-            text("ALTER TABLE llm_relationship_note ADD COLUMN affinity DOUBLE PRECISION NOT NULL DEFAULT 0")
-        )
-    if "rage" not in names:
-        connection.execute(text("ALTER TABLE llm_relationship_note ADD COLUMN rage INTEGER NOT NULL DEFAULT 0"))
-    if "rage_last_attack_at" not in names:
-        connection.execute(
-            text("ALTER TABLE llm_relationship_note ADD COLUMN rage_last_attack_at BIGINT NOT NULL DEFAULT 0")
-        )
-    if "rage_last_attack_message_id" not in names:
-        connection.execute(
-            text("ALTER TABLE llm_relationship_note ADD COLUMN rage_last_attack_message_id BIGINT NOT NULL DEFAULT 0")
-        )
-    if "rage_silenced_until" not in names:
-        connection.execute(
-            text("ALTER TABLE llm_relationship_note ADD COLUMN rage_silenced_until BIGINT NOT NULL DEFAULT 0")
-        )
-    if "rage_silence_reason" not in names:
-        connection.execute(
-            text("ALTER TABLE llm_relationship_note ADD COLUMN rage_silence_reason TEXT NOT NULL DEFAULT ''")
-        )
-
-
-def _ensure_pg_llm_memory_embedding_columns(connection) -> None:
-    """旧库 llm_memory_entry 缺 embedding 缓存列时补齐。"""
-    insp = inspect(connection)
-    if not insp.has_table("llm_memory_entry"):
-        return
-    names = {c["name"] for c in insp.get_columns("llm_memory_entry")}
-    if "embedding_json" not in names:
-        connection.execute(text("ALTER TABLE llm_memory_entry ADD COLUMN embedding_json TEXT"))
-    if "embedding_model" not in names:
-        connection.execute(text("ALTER TABLE llm_memory_entry ADD COLUMN embedding_model TEXT"))
-
-
-def _ensure_pg_llm_memory_metadata_columns(connection) -> None:
-    """旧库 llm_memory_entry 缺治理元数据列时补齐。"""
-    insp = inspect(connection)
-    if not insp.has_table("llm_memory_entry"):
-        return
-    names = {c["name"] for c in insp.get_columns("llm_memory_entry")}
-    if "importance" not in names:
-        connection.execute(
-            text("ALTER TABLE llm_memory_entry ADD COLUMN importance DOUBLE PRECISION NOT NULL DEFAULT 0.5")
-        )
-    if "confidence" not in names:
-        connection.execute(
-            text("ALTER TABLE llm_memory_entry ADD COLUMN confidence DOUBLE PRECISION NOT NULL DEFAULT 0.5")
-        )
-    if "expires_at" not in names:
-        connection.execute(text("ALTER TABLE llm_memory_entry ADD COLUMN expires_at BIGINT NOT NULL DEFAULT 0"))
-    if "visibility" not in names:
-        connection.execute(text("ALTER TABLE llm_memory_entry ADD COLUMN visibility TEXT NOT NULL DEFAULT 'private'"))
-
-
-def _ensure_pg_llm_memory_graph_columns(connection) -> None:
-    """旧库记忆图谱表补 soft-delete 列。"""
-    insp = inspect(connection)
-    if insp.has_table("llm_memory_entity"):
-        names = {c["name"] for c in insp.get_columns("llm_memory_entity")}
-        if "deleted_at" not in names:
-            connection.execute(text("ALTER TABLE llm_memory_entity ADD COLUMN deleted_at BIGINT"))
-
-
-def _ensure_pg_message_group_time_index(connection) -> None:
-    """message 表补 group_id+time 索引。"""
-    insp = inspect(connection)
-    if not insp.has_table("message"):
-        return
-    connection.execute(text("CREATE INDEX IF NOT EXISTS ix_message_group_time ON message (group_id, time)"))
-
-
-def _ensure_pg_message_group_user_time_index(connection) -> None:
-    """message 表补 group_id+user_id+time 索引。"""
-    insp = inspect(connection)
-    if not insp.has_table("message"):
-        return
-    connection.execute(
-        text("CREATE INDEX IF NOT EXISTS ix_message_group_user_time ON message (group_id, user_id, time)")
-    )
-
-
-def _ensure_pg_message_timeline_metadata(connection) -> None:
-    """旧 message 表补群时间线所需的身份与引用元数据。"""
-    insp = inspect(connection)
-    if not insp.has_table("message"):
-        return
-    names = {column["name"] for column in insp.get_columns("message")}
-    if "sender_name" not in names:
-        connection.execute(text("ALTER TABLE message ADD COLUMN sender_name TEXT NOT NULL DEFAULT ''"))
-    if "message_id" not in names:
-        connection.execute(text("ALTER TABLE message ADD COLUMN message_id BIGINT"))
-    if "reply_to_message_id" not in names:
-        connection.execute(text("ALTER TABLE message ADD COLUMN reply_to_message_id BIGINT"))
-    if "suppressed_by_rage" not in names:
-        connection.execute(text("ALTER TABLE message ADD COLUMN suppressed_by_rage BOOLEAN NOT NULL DEFAULT FALSE"))
-
-
-def _ensure_pg_message_unique_anchor(connection) -> None:
-    """message 表补 (group_id, bot_id, message_id) 唯一锚点，落库幂等。
-
-    历史数据在同锚点下可能已重复，先删多余行（保留最小 id）再建唯一约束。
-    message_id 为 NULL 的行不参与冲突（PG 中 NULL <> NULL），不受影响。
-    """
-    insp = inspect(connection)
-    if not insp.has_table("message"):
-        return
-    connection.execute(
-        text(
-            "DELETE FROM message a USING message b "
-            "WHERE a.message_id IS NOT NULL "
-            "AND b.message_id IS NOT NULL "
-            "AND a.group_id = b.group_id "
-            "AND a.bot_id = b.bot_id "
-            "AND a.message_id = b.message_id "
-            "AND a.id > b.id"
-        )
-    )
-    connection.execute(
-        text(
-            "CREATE UNIQUE INDEX IF NOT EXISTS uq_message_group_bot_message_id "
-            "ON message (group_id, bot_id, message_id)"
-        )
-    )
-
-
-def _ensure_pg_context_answer_reply_index(connection) -> None:
-    """context_answer 表补 context_id+count+time 索引。"""
-    insp = inspect(connection)
-    if not insp.has_table("context_answer"):
-        return
-    connection.execute(text("DROP INDEX IF EXISTS ix_context_answer_context_id"))
-    connection.execute(
-        text("CREATE INDEX IF NOT EXISTS ix_context_answer_ctx_count_time ON context_answer (context_id, count, time)")
-    )
-
-
-def _ensure_pg_context_answer_message_reply_index(connection) -> None:
-    """context_answer_message 表补 answer_id+id 索引。"""
-    insp = inspect(connection)
-    if not insp.has_table("context_answer_message"):
-        return
-    connection.execute(
-        text(
-            "CREATE INDEX IF NOT EXISTS ix_context_answer_message_answer_id_id "
-            "ON context_answer_message (answer_id, id)"
-        )
-    )
-
-
-def _ensure_pg_background_job_delivery_claim_index(connection) -> None:
-    """background_job 表补已完成视觉投递领取索引。"""
-    insp = inspect(connection)
-    if not insp.has_table("background_job"):
-        return
-    connection.execute(
-        text(
-            "CREATE INDEX IF NOT EXISTS ix_background_job_delivery_claim ON background_job (kind, status, finished_at)"
-        )
-    )
-
-
-def _ensure_pg_background_job_pending_claim_index(connection) -> None:
-    """background_job 表补 pending 领取部分索引，避免 claim 扫全部历史 available 行。"""
-    insp = inspect(connection)
-    if not insp.has_table("background_job"):
-        return
-    connection.execute(
-        text(
-            "CREATE INDEX IF NOT EXISTS ix_background_job_pending_claim "
-            "ON background_job (available_at, created_at) "
-            "WHERE status = 'pending' AND finished_at IS NULL"
-        )
-    )
-
-
-def _ensure_pg_stat_statements_extension(connection) -> None:
-    """启用 pg_stat_statements（仅应在独立 autocommit 连接中调用）。
-
-    勿放在 schema 迁移同一事务内：非超级用户失败会污染事务，导致整次 init 回滚。
-    """
-    connection.execute(text("CREATE EXTENSION IF NOT EXISTS pg_stat_statements"))
-
-
-async def try_enable_pg_stat_statements(engine: AsyncEngine) -> bool:
-    """可选诊断扩展：独立短事务，失败回滚本事务且不阻断启动。"""
-    try:
-        async with engine.begin() as conn:
-            await conn.run_sync(_ensure_pg_stat_statements_extension)
-        return True
-    except Exception as exc:
-        logger.warning(
-            "数据库：跳过 pg_stat_statements（需扩展权限或 shared_preload_libraries），诊断将降级。详情: {}",
-            exc,
-        )
-        return False
-
-
-def is_pg_initialized() -> bool:
-    return _session_factory is not None
-
-
-def pg_engine() -> AsyncEngine | None:
-    return _engine
-
-
-def pg_pool_live_stats() -> dict[str, int] | None:
-    """运行时连接池占用。"""
-    if _engine is None:
-        return None
-    from pallas.core.foundation.db.pool_budget import pg_pool_capacity
-
-    pool = _engine.pool
-    return {
-        "pool_size": int(pool.size()),
-        "checked_out": int(pool.checkedout()),
-        "overflow": int(pool.overflow()),
-        "capacity": pg_pool_capacity(),
-    }
-
-
-@asynccontextmanager
-async def get_session(*, read_only: bool = False):
-    if _session_factory is None:
-        raise RuntimeError("PostgreSQL 尚未初始化，请先调用 init_pg()")
-    from pallas.core.foundation.db.pool_diagnostics import (
-        note_slow_pg_session,
-        pg_session_caller_hint_entry,
-        session_hold_warn_ms,
-    )
-
-    caller = pg_session_caller_hint_entry()
-    session = _session_factory()
-    # 只读路径用 AUTOCOMMIT：多段 SELECT 之间不会长期占着 idle in transaction 连接。
-    if read_only:
-        await session.connection(execution_options={"isolation_level": "AUTOCOMMIT"})
-    t0 = time.monotonic()
-    try:
-        yield session
-    except BaseException:
-        # CancelledError 非 Exception 子类；清理须 shield，避免 close/rollback 再被取消导致连接未归还池
-        if not read_only:
-            with contextlib.suppress(BaseException):
-                await asyncio.shield(session.rollback())
-        raise
-    finally:
-        held_ms = (time.monotonic() - t0) * 1000.0
-        if held_ms >= session_hold_warn_ms():
-            note_slow_pg_session(held_ms, caller)
-        try:
-            if not read_only:
-                with contextlib.suppress(BaseException):
-                    await asyncio.shield(session.rollback())
-            await asyncio.shield(session.close())
-        except BaseException:
-            with contextlib.suppress(BaseException):
-                await asyncio.shield(session.invalidate())
-
-
-# 启动期 DDL ensure 注册表（step_id 稳定，供控制台与可观测使用）
-PG_SCHEMA_ENSURE_STEPS: list[tuple[str, Any]] = [
-    ("ddl.image_cache_blob_data", _ensure_pg_image_cache_blob_data),
-    ("ddl.image_cache_content_hash", _ensure_pg_image_cache_content_hash),
-    ("ddl.image_cache_blob_path", _ensure_pg_image_cache_blob_path),
-    ("ddl.group_config_blocked_user_ids", _ensure_pg_group_config_blocked_user_ids),
-    ("ddl.background_job_lease_id", _ensure_pg_background_job_lease_id),
-    ("ddl.group_config_style_profile", _ensure_pg_group_config_style_profile),
-    ("ddl.group_config_plugin_storage", _ensure_pg_group_config_plugin_storage),
-    ("ddl.group_config_disabled_plugins_audit", _ensure_pg_group_config_disabled_plugins_audit),
-    ("ddl.bot_config_disabled_plugins_audit", _ensure_pg_bot_config_disabled_plugins_audit),
-    ("ddl.bot_config_community_roster_show_qq", _ensure_pg_bot_config_community_roster_show_qq),
-    ("ddl.bot_config_group_style_enabled", _ensure_pg_bot_config_group_style_enabled),
-    ("ddl.bot_config_persona", _ensure_pg_bot_config_persona),
-    ("ddl.bot_config_plugin_storage", _ensure_pg_bot_config_plugin_storage),
-    ("ddl.user_config_maa_devices", _ensure_pg_user_config_maa_devices),
-    ("ddl.user_config_ban_audit", _ensure_pg_user_config_ban_audit),
-    ("ddl.llm_memory_embedding_columns", _ensure_pg_llm_memory_embedding_columns),
-    ("ddl.llm_memory_metadata_columns", _ensure_pg_llm_memory_metadata_columns),
-    ("ddl.llm_memory_graph_columns", _ensure_pg_llm_memory_graph_columns),
-    ("ddl.llm_relationship_delta_columns", _ensure_pg_llm_relationship_delta_columns),
-    ("ddl.message_group_time_index", _ensure_pg_message_group_time_index),
-    ("ddl.message_group_user_time_index", _ensure_pg_message_group_user_time_index),
-    ("ddl.message_timeline_metadata", _ensure_pg_message_timeline_metadata),
-    ("ddl.message_unique_anchor", _ensure_pg_message_unique_anchor),
-    ("ddl.context_answer_reply_index", _ensure_pg_context_answer_reply_index),
-    ("ddl.context_answer_message_reply_index", _ensure_pg_context_answer_message_reply_index),
-    ("ddl.background_job_delivery_claim_index", _ensure_pg_background_job_delivery_claim_index),
-    ("ddl.background_job_pending_claim_index", _ensure_pg_background_job_pending_claim_index),
-]
-
-
-async def init_pg(engine: AsyncEngine) -> None:
-    """创建表结构并注入 engine；对已有 PG 库补全 group_config.blocked_user_ids 等轻量迁移。"""
-    global _engine, _session_factory
-    from pallas.core.foundation.db.schema_registry import run_registered_pg_ensures
-
-    _engine = engine
-    _session_factory = async_sessionmaker(engine, expire_on_commit=False)
-
-    async with engine.begin() as conn:
-        await conn.run_sync(Base.metadata.create_all)
-        await conn.run_sync(run_registered_pg_ensures)
-
-
-async def dispose_pg() -> None:
-    """关闭连接池并清空配置 TTL 缓存，bot 退出或测试 teardown 时调用。"""
-    global _engine, _session_factory
-    if _engine is not None:
-        await _engine.dispose()
-        _engine = None
-    # 释放 engine 后清空 session factory
-    _session_factory = None
-    await clear_reply_query_snapshot_cache(None)
-    # schema 重建后清空 ORM 缓存
-    for cache in _CONFIG_CACHES.values():
-        await cache.clear()
-
-
-_LOAD_RELATED = [
-    selectinload(ContextRow.answers).selectinload(ContextAnswerRow.messages),
-    selectinload(ContextRow.ban),
-]
-
-_LOAD_REPLY_CTX = [
-    selectinload(ContextRow.ban),
-]
-
-
-def keywords_hash(keywords: str) -> str:
-    # 先剥除 \x00 再哈希，与 ContextRow.keywords 实际存储值保持一致
-    clean = keywords.replace("\x00", "") if keywords and "\x00" in keywords else keywords
-    return hashlib.md5((clean or "").encode("utf-8", errors="replace")).hexdigest()
-
-
-# asyncpg 单语句参数上限 32767
-_ANSWER_BATCH = 500  # ContextAnswerRow 6 列 × 500 = 3000
-_MSG_BATCH = 16000  # ContextAnswerMessageRow 2 列 × 16000 = 32000
-
-
-async def delete_context_answer_orphans(
-    session: AsyncSession,
-    *,
-    ctx_id: int,
-    kept_ids: list[int],
-    chunk_size: int = _DELETE_ID_BATCH,
-) -> None:
-    """删除指定 Context 下未保留的 Answer，避免生成超长 NOT IN 参数列表。"""
-    if not kept_ids:
-        await session.execute(delete(ContextAnswerRow).where(ContextAnswerRow.context_id == ctx_id))
-        return
-
-    existing_ids = (
-        (await session.execute(select(ContextAnswerRow.id).where(ContextAnswerRow.context_id == ctx_id)))
-        .scalars()
-        .all()
-    )
-    kept_id_set = set(kept_ids)
-    orphan_ids = [int(ans_id) for ans_id in existing_ids if int(ans_id) not in kept_id_set]
-    for offset in range(0, len(orphan_ids), chunk_size):
-        chunk = orphan_ids[offset : offset + chunk_size]
-        await session.execute(
-            delete(ContextAnswerRow).where(ContextAnswerRow.context_id == ctx_id, ContextAnswerRow.id.in_(chunk))
-        )
-
-
-_BAN_BATCH = 6000  # ContextBanRow 5 列 × 6000 = 30000
-
-
-async def _insert_answers_batched(session: AsyncSession, context_id: int, answers) -> None:
-    """分批插入 ContextAnswerRow 及其关联的 ContextAnswerMessageRow"""
-
-    for i in range(0, len(answers), _ANSWER_BATCH):
-        batch: list[Answer] = answers[i : i + _ANSWER_BATCH]
-        rows = []
-        for a in batch:
-            kw = _s(a.keywords) or ""
-            rows.append(
-                ContextAnswerRow(
-                    context_id=context_id,
-                    keywords=kw,
-                    keywords_hash=keywords_hash(kw),
-                    group_id=a.group_id,
-                    count=a.count,
-                    time=a.time,
-                )
-            )
-        session.add_all(rows)
-        await session.flush()
-
-        msg_rows = [
-            ContextAnswerMessageRow(answer_id=rows[j].id, message=_s(m) or "")
-            for j, a in enumerate(batch)
-            for m in a.messages
-        ]
-        for k in range(0, len(msg_rows), _MSG_BATCH):
-            session.add_all(msg_rows[k : k + _MSG_BATCH])
-            await session.flush()
-
-
-async def _insert_bans_batched(session: AsyncSession, context_id: int, bans) -> None:
-    """分批插入 ContextBanRow"""
-    for i in range(0, len(bans), _BAN_BATCH):
-        batch = bans[i : i + _BAN_BATCH]
-        session.add_all([
-            ContextBanRow(
-                context_id=context_id,
-                keywords=_s(b.keywords) or "",
-                group_id=b.group_id,
-                reason=_s(b.reason) or "",
-                time=b.time,
-            )
-            for b in batch
-        ])
-        await session.flush()
-
-
-def row_to_context(row: ContextRow, *, reply_messages: dict[int, list[str]] | None = None) -> Context:
-    from pallas.core.foundation.db.modules import Answer, Ban, Context
-
-    answers = []
-    for a in row.answers:
-        if reply_messages is not None:
-            msgs = list(reply_messages.get(int(a.id), []))
-        else:
-            msgs = [m.message for m in a.messages]
-        answers.append(
-            Answer.model_construct(
-                keywords=a.keywords,
-                group_id=a.group_id,
-                count=a.count,
-                time=a.time,
-                messages=msgs,
-            )
-        )
-    ban = [
-        Ban.model_construct(
-            keywords=b.keywords,
-            group_id=b.group_id,
-            reason=b.reason,
-            time=b.time,
-        )
-        for b in row.ban
-    ]
-    return Context.model_construct(
-        keywords=row.keywords,
-        time=row.time,
-        trigger_count=row.trigger_count,
-        answers=answers,
-        ban=ban,
-        clear_time=row.clear_time,
-    )
-
-
-def build_reply_context(
-    *,
-    keywords: str,
-    time_value: int,
-    trigger_count: int,
-    clear_time: int,
-    answer_rows: list[ContextAnswerRow],
-    ban_rows: list[ContextBanRow],
-    reply_messages: dict[int, list[str]],
-):
-    from pallas.core.foundation.db.modules import Answer, Ban, Context
-
-    return Context.model_construct(
-        keywords=keywords,
-        time=time_value,
-        trigger_count=trigger_count,
-        answers=[
-            Answer.model_construct(
-                keywords=answer.keywords,
-                group_id=answer.group_id,
-                count=answer.count,
-                time=answer.time,
-                messages=list(reply_messages.get(int(answer.id), [])),
-            )
-            for answer in answer_rows
-        ],
-        ban=[
-            Ban.model_construct(
-                keywords=ban.keywords,
-                group_id=ban.group_id,
-                reason=ban.reason,
-                time=ban.time,
-            )
-            for ban in ban_rows
-        ],
-        clear_time=clear_time,
-    )
-
-
-def build_reply_message_query(answer_ids: list[int], msg_cap: int):
-    rn = (
-        func
-        .row_number()
-        .over(
-            partition_by=ContextAnswerMessageRow.answer_id,
-            order_by=ContextAnswerMessageRow.id.desc(),
-        )
-        .label("rn")
-    )
-    ranked = (
-        select(
-            ContextAnswerMessageRow.answer_id,
-            ContextAnswerMessageRow.message,
-            ContextAnswerMessageRow.id,
-            rn,
-        )
-        .where(ContextAnswerMessageRow.answer_id.in_(answer_ids))
-        .subquery()
-    )
-    return (
-        select(ranked.c.answer_id, ranked.c.message, ranked.c.id)
-        .where(ranked.c.rn <= msg_cap)
-        .order_by(ranked.c.answer_id, ranked.c.id)
-    )
-
-
-async def clear_reply_query_snapshot_cache(keywords: str | None = None) -> None:
-    async with _reply_query_snapshot_lock:
-        if keywords is None:
-            _reply_query_snapshot_cache.clear()
-            _reply_query_snapshot_inflight.clear()
-            return
-        key = keywords.strip()
-        if key:
-            _reply_query_snapshot_cache.pop(key, None)
-
-
-async def cached_reply_query_snapshot(
-    keywords: str,
-    loader,
-) -> Context | None:
-    from pallas.core.foundation.db.pool_budget import is_pg_pool_timeout_error, pg_pool_under_pressure
-    from pallas.core.platform.ingress.hotpath_metrics import record_reply_snapshot
-    from pallas.product.corpus.find_cache import mark_reply_db_fail, reply_db_fail_active
-    from pallas.product.corpus.reply_perf_config import (
-        reply_snapshot_max_entries,
-        reply_snapshot_query_timeout_sec,
-        reply_snapshot_ttl_sec,
-    )
-
-    key = (keywords or "").strip()
-    if not key:
-        return None
-    if pg_pool_under_pressure(threshold=0.55):
-        record_reply_snapshot(hit=False, skipped=True)
-        logger.debug(
-            "Reply query snapshot skipped because the PostgreSQL pool is under pressure for keyword length [{}].",
-            len(key),
-        )
-        return None
-    if reply_db_fail_active(key):
-        record_reply_snapshot(hit=False, skipped=True)
-        logger.debug(
-            "Reply query snapshot skipped during database-failure cooldown for keyword length [{}].",
-            len(key),
-        )
-        return None
-    now = time.monotonic()
-    task: asyncio.Task[Context | None] | None = None
-    async with _reply_query_snapshot_lock:
-        hit = _reply_query_snapshot_cache.get(key)
-        if hit is not None:
-            expire_at, value = hit
-            if now < expire_at:
-                _reply_query_snapshot_cache.move_to_end(key)
-                record_reply_snapshot(hit=True)
-                return value
-            _reply_query_snapshot_cache.pop(key, None)
-        task = _reply_query_snapshot_inflight.get(key)
-        if task is None:
-            task = asyncio.create_task(loader(key))
-            _reply_query_snapshot_inflight[key] = task
-            record_reply_snapshot(hit=False)
-        else:
-            # 与 inflight 合并：算命中同类请求，避免重复打库计数偏高
-            record_reply_snapshot(hit=True)
-
-    try:
-        ctx = await asyncio.wait_for(asyncio.shield(task), timeout=reply_snapshot_query_timeout_sec())
-    except TimeoutError:
-        async with _reply_query_snapshot_lock:
-            if _reply_query_snapshot_inflight.get(key) is task:
-                _reply_query_snapshot_inflight.pop(key, None)
-        record_reply_snapshot(hit=False)
-        logger.debug("Reply query snapshot timed out for keyword length [{}].", len(key))
-        return None
-    except Exception as exc:
-        async with _reply_query_snapshot_lock:
-            if _reply_query_snapshot_inflight.get(key) is task:
-                _reply_query_snapshot_inflight.pop(key, None)
-        if is_pg_pool_timeout_error(exc):
-            mark_reply_db_fail(key)
-            logger.debug(
-                "Reply query snapshot skipped after a database timeout for keyword length [{}].",
-                len(key),
-            )
-            return None
-        raise
-
-    async with _reply_query_snapshot_lock:
-        if _reply_query_snapshot_inflight.get(key) is task:
-            _reply_query_snapshot_inflight.pop(key, None)
-        _reply_query_snapshot_cache[key] = (time.monotonic() + reply_snapshot_ttl_sec(), ctx)
-        _reply_query_snapshot_cache.move_to_end(key)
-        while len(_reply_query_snapshot_cache) > reply_snapshot_max_entries():
-            _reply_query_snapshot_cache.popitem(last=False)
-    return ctx
-
-
-def row_to_blacklist(row: BlackListRow):
-    from pallas.core.foundation.db.modules import BlackList
-
-    return BlackList.model_construct(
-        group_id=row.group_id,
-        answers=list(row.answers),
-        answers_reserve=list(row.answers_reserve),
-    )
-
-
-def row_to_image_cache(row: ImageCacheRow) -> ImageCache:
-    from pallas.core.foundation.db.modules import ImageCache
-
-    return ImageCache.model_construct(
-        cq_code=row.cq_code,
-        content_hash=row.content_hash,
-        blob_data=row.blob_data,
-        blob_path=row.blob_path,
-        blob_size=row.blob_size,
-        ref_times=row.ref_times,
-        date=row.date,
-    )
-
-
-class PgContextRepository:
-    async def context_exists_by_keywords(self, keywords: str) -> bool:
-        khash = keywords_hash(keywords)
-        async with get_session(read_only=True) as session:
-            result = await session.execute(select(ContextRow.id).where(ContextRow.keywords_hash == khash).limit(1))
-            return result.scalar_one_or_none() is not None
-
-    async def find_by_keywords(self, keywords: str) -> Context | None:
-        khash = keywords_hash(keywords)
-        async with get_session(read_only=True) as session:
-            result = await session.execute(
-                select(ContextRow).options(*_LOAD_RELATED).where(ContextRow.keywords_hash == khash)
-            )
-            row = result.scalar_one_or_none()
-            return row_to_context(row) if row else None
-
-    async def find_by_keywords_for_reply(self, keywords: str) -> Context | None:
-        return await cached_reply_query_snapshot(keywords, self._find_by_keywords_for_reply_uncached)
-
-    async def _find_by_keywords_for_reply_uncached(self, keywords: str) -> Context | None:
-        """接话路径：轻量列查询 + 限量 Answer/Message，避免 ORM 关联热路径放大。"""
-        return await self._find_by_keywords_for_reply_snapshot(keywords)
-
-    async def _find_by_keywords_for_reply_snapshot(self, keywords: str) -> Context | None:
-        """一次受限快照读取接话所需的 context、ban、answer 与 message。"""
-        from pallas.core.foundation.db.modules import Answer, Ban, Context
-
-        khash = keywords_hash(keywords)
-        from pallas.product.corpus.reply_perf_config import reply_query_caps
-
-        msg_cap, ans_cap = reply_query_caps(keywords)
-        t_start = time.monotonic()
-        async with get_session(read_only=True) as session:
-            context_row = (
-                (
-                    await session.execute(
-                        text(
-                            """
-                        SELECT c.id, c.keywords, c.time, c.trigger_count, c.clear_time,
-                            COALESCE((
-                                SELECT jsonb_agg(jsonb_build_object(
-                                    'keywords', b.keywords, 'group_id', b.group_id,
-                                    'reason', b.reason, 'time', b.time
-                                ) ORDER BY b.id)
-                                FROM context_ban b WHERE b.context_id = c.id
-                            ), '[]'::jsonb) AS bans
-                        FROM context c WHERE c.keywords_hash = :khash
-                        """
-                        ),
-                        {"khash": khash},
-                    )
-                )
-                .mappings()
-                .one_or_none()
-            )
-            row = None
-            if context_row is not None:
-                answer_row = (
-                    (
-                        await session.execute(
-                            text(
-                                """
-                            WITH answers AS (
-                                SELECT a.id, a.keywords, a.group_id, a.count, a.time
-                                FROM context_answer a
-                                WHERE a.context_id = :context_id
-                                ORDER BY a.count DESC, a.time DESC LIMIT :ans_cap
-                            ), ranked_messages AS (
-                                SELECT m.answer_id, m.message, m.id,
-                                       row_number() OVER (PARTITION BY m.answer_id ORDER BY m.id DESC) AS rn
-                                FROM context_answer_message m JOIN answers a ON a.id = m.answer_id
-                            ), messages AS (
-                                SELECT answer_id, jsonb_agg(message ORDER BY id) AS values
-                                FROM ranked_messages WHERE rn <= :msg_cap GROUP BY answer_id
-                            )
-                            SELECT COALESCE(jsonb_agg(jsonb_build_object(
-                                'keywords', a.keywords, 'group_id', a.group_id,
-                                'count', a.count, 'time', a.time,
-                                'messages', COALESCE(m.values, '[]'::jsonb)
-                            ) ORDER BY a.count DESC, a.time DESC), '[]'::jsonb) AS answers
-                            FROM answers a LEFT JOIN messages m ON m.answer_id = a.id
-                            """
-                            ),
-                            {"context_id": context_row["id"], "ans_cap": ans_cap, "msg_cap": msg_cap},
-                        )
-                    )
-                    .mappings()
-                    .one()
-                )
-                row = {**context_row, "answers": answer_row["answers"]}
-        elapsed_ms = (time.monotonic() - t_start) * 1000.0
-        if row is None:
-            self._log_reply_query_slow(
-                keywords=keywords,
-                elapsed_ms=elapsed_ms,
-                context_ms=None,
-                ban_ms=None,
-                answer_ms=None,
-                message_ms=None,
-                ban_count=0,
-                answer_count=0,
-                message_count=0,
-                hit=False,
-            )
-            return None
-        bans = row["bans"] if isinstance(row["bans"], list) else json.loads(row["bans"])
-        answers = row["answers"] if isinstance(row["answers"], list) else json.loads(row["answers"])
-        self._log_reply_query_slow(
-            keywords=keywords,
-            elapsed_ms=elapsed_ms,
-            context_ms=None,
-            ban_ms=None,
-            answer_ms=None,
-            message_ms=None,
-            ban_count=len(bans),
-            answer_count=len(answers),
-            message_count=sum(len(answer["messages"]) for answer in answers),
-            hit=True,
-        )
-        return Context.model_construct(
-            keywords=row["keywords"],
-            time=row["time"],
-            trigger_count=row["trigger_count"],
-            clear_time=row["clear_time"],
-            ban=[Ban.model_construct(**ban) for ban in bans],
-            answers=[Answer.model_construct(**answer) for answer in answers],
-        )
-
-    @staticmethod
-    def _log_reply_query_slow(
-        *,
-        keywords: str,
-        elapsed_ms: float,
-        context_ms: float | None,
-        ban_ms: float | None,
-        answer_ms: float | None,
-        message_ms: float | None,
-        ban_count: int,
-        answer_count: int,
-        message_count: int,
-        hit: bool,
-    ) -> None:
-        from pallas.core.platform.ingress.hotpath_metrics import record_reply_query_stages
-
-        record_reply_query_stages(
-            context_ms=context_ms,
-            ban_ms=ban_ms,
-            answer_ms=answer_ms,
-            message_ms=message_ms,
-            total_ms=elapsed_ms,
-        )
-        threshold_ms = slow_path_threshold_ms("PALLAS_SLOW_REPLY_QUERY_MS", 250.0)
-        if elapsed_ms < threshold_ms:
-            return
-        logger.debug(
-            "Corpus reply query used the slow path in [{:.1f}]ms: context [{:.1f}]ms, ban [{:.1f}]ms, "
-            "answers [{:.1f}]ms, messages [{:.1f}]ms; counts were ban [{}], answers [{}], "
-            "messages [{}], hit [{}], and keyword length [{}].",
-            elapsed_ms,
-            context_ms or 0.0,
-            ban_ms or 0.0,
-            answer_ms or 0.0,
-            message_ms or 0.0,
-            ban_count,
-            answer_count,
-            message_count,
-            hit,
-            len(keywords),
-        )
-
-    async def save(self, context: Context) -> None:
-        khash = keywords_hash(context.keywords)
-        async with get_session() as session:
-            result = await session.execute(select(ContextRow).where(ContextRow.keywords_hash == khash))
-            row = result.scalar_one_or_none()
-
-            if row is None:
-                row = ContextRow(
-                    keywords=_s(context.keywords) or "",
-                    keywords_hash=khash,
-                    time=context.time,
-                    trigger_count=context.trigger_count,
-                    clear_time=context.clear_time,
-                )
-                session.add(row)
-                await session.flush()
-            else:
-                row.time = context.time
-                row.trigger_count = context.trigger_count
-                row.clear_time = context.clear_time
-                await session.execute(delete(ContextAnswerRow).where(ContextAnswerRow.context_id == row.id))
-                await session.execute(delete(ContextBanRow).where(ContextBanRow.context_id == row.id))
-
-            await _insert_answers_batched(session, row.id, context.answers)
-            await _insert_bans_batched(session, row.id, context.ban)
-            await session.commit()
-        await clear_reply_query_snapshot_cache(context.keywords)
-
-    async def insert(self, context: Context) -> None:
-        """插入新 Context。并发下同 keywords 第二个写入会被 unique 约束拒绝，等价为 no-op。"""
-        khash = keywords_hash(context.keywords)
-        try:
-            async with get_session() as session:
-                row = ContextRow(
-                    keywords=_s(context.keywords) or "",
-                    keywords_hash=khash,
-                    time=context.time,
-                    trigger_count=context.trigger_count,
-                    clear_time=context.clear_time,
-                )
-                session.add(row)
-                await session.flush()
-                await _insert_answers_batched(session, row.id, context.answers)
-                await _insert_bans_batched(session, row.id, context.ban)
-                await session.commit()
-            await clear_reply_query_snapshot_cache(context.keywords)
-        except IntegrityError:
-            pass
-
-    _DELETE_EXPIRED_CHUNK = 10000
-
-    async def delete_expired(self, expiration: int, threshold: int) -> None:
-        """分批删除过期 Context，避免千万级时长锁表。级联删除由 FK ondelete=CASCADE 处理。"""
-        deleted_any = False
-        while True:
-            async with get_session() as session:
-                subq = (
-                    select(ContextRow.id)
-                    .where(ContextRow.time < expiration, ContextRow.trigger_count < threshold)
-                    .limit(self._DELETE_EXPIRED_CHUNK)
-                    .subquery()
-                )
-                result = await session.execute(
-                    delete(ContextRow).where(ContextRow.id.in_(select(subq.c.id))).returning(ContextRow.id)
-                )
-                deleted = len(result.scalars().all())
-                await session.commit()
-            deleted_any = deleted_any or deleted > 0
-            if deleted < self._DELETE_EXPIRED_CHUNK:
-                break
-        if deleted_any:
-            await clear_reply_query_snapshot_cache(None)
-
-    _CLEANUP_CHUNK = 500
-
-    async def find_for_cleanup(self, trigger_threshold: int, expiration: int) -> list[Context]:
-        """
-        语义对齐 Mongo：trigger_count > threshold OR clear_time < expiration。
-        流式按主键 id 分页，避免千万级时一次性全加载 OOM。
-        """
-        results: list[Context] = []
-        last_id = 0
-        while True:
-            async with get_session(read_only=True) as session:
-                result = await session.execute(
-                    select(ContextRow)
-                    .options(*_LOAD_RELATED)
-                    .where(
-                        or_(
-                            ContextRow.trigger_count > trigger_threshold,
-                            ContextRow.clear_time < expiration,
-                        ),
-                        ContextRow.id > last_id,
-                    )
-                    .order_by(ContextRow.id)
-                    .limit(self._CLEANUP_CHUNK)
-                )
-                rows = list(result.scalars().all())
-            if not rows:
-                break
-            results.extend(row_to_context(r) for r in rows)
-            last_id = rows[-1].id
-            if len(rows) < self._CLEANUP_CHUNK:
-                break
-        return results
-
-    async def upsert_answer(
-        self,
-        keywords: str,
-        group_id: int,
-        answer_keywords: str,
-        answer_time: int,
-        message: str,
-        append_on_existing: bool,
-    ) -> None:
-        """
-        原子 upsert，依赖 UNIQUE(context_id, group_id, keywords)：
-          - INSERT ... ON CONFLICT DO UPDATE SET count = count + 1, time = EXCLUDED.time
-          - RETURNING 中借助 xmax 判断 insert vs update，决定是否 append message
-          - 最后原子递增 Context.trigger_count / 更新 time
-        """
-        khash = keywords_hash(keywords)
-        ans_kw_s = _s(answer_keywords) or ""
-        msg_s = _s(message) or ""
-        from pallas.product.llm.corpus_contamination import reject_corpus_learn_message
-
-        if reject_corpus_learn_message(msg_s, source="upsert_answer"):
-            return
-
-        async with get_session() as session:
-            ctx_result = await session.execute(select(ContextRow.id).where(ContextRow.keywords_hash == khash))
-            ctx_id = ctx_result.scalar_one_or_none()
-            if ctx_id is None:
-                return
-
-            stmt = pg_insert(ContextAnswerRow).values(
-                context_id=ctx_id,
-                keywords=ans_kw_s,
-                keywords_hash=keywords_hash(ans_kw_s),
-                group_id=group_id,
-                count=1,
-                time=answer_time,
-            )
-            stmt = stmt.on_conflict_do_update(
-                constraint="uq_context_answer_ctx_group_kw",
-                set_={
-                    "count": ContextAnswerRow.count + 1,
-                    "time": stmt.excluded.time,
-                },
-            ).returning(ContextAnswerRow.id, literal_column("(xmax = 0)").label("was_insert"))
-
-            row = (await session.execute(stmt)).first()
-            assert row is not None
-            ans_id, was_insert = int(row.id), bool(row.was_insert)
-
-            if was_insert or append_on_existing:
-                await session.execute(insert(ContextAnswerMessageRow).values(answer_id=ans_id, message=msg_s))
-
-            await session.execute(
-                update(ContextRow)
-                .where(ContextRow.id == ctx_id)
-                .values(trigger_count=ContextRow.trigger_count + 1, time=answer_time)
-            )
-            await session.commit()
-
-    async def learn_answer(
-        self,
-        *,
-        keywords: str,
-        group_id: int,
-        answer_keywords: str,
-        answer_time: int,
-        message: str,
-        append_on_existing: bool,
-    ) -> bool:
-        """
-        学习热路径专用：
-          - Context 不存在时直接原子创建并写入首条 Answer
-          - Context 已存在时在同一事务内原子 upsert Answer
-        返回值表示本次是否新建了 Context。
-        """
-        khash = keywords_hash(keywords)
-        kw_s = _s(keywords) or ""
-        ans_kw_s = _s(answer_keywords) or ""
-        msg_s = _s(message) or ""
-        from pallas.product.llm.corpus_contamination import reject_corpus_learn_message
-
-        if reject_corpus_learn_message(msg_s, source="learn_answer"):
-            return False
-
-        async with get_session() as session:
-            ctx_stmt = pg_insert(ContextRow).values(
-                keywords=kw_s,
-                keywords_hash=khash,
-                time=answer_time,
-                trigger_count=1,
-                clear_time=0,
-            )
-            ctx_stmt = ctx_stmt.on_conflict_do_update(
-                index_elements=[ContextRow.keywords_hash],
-                set_={
-                    "trigger_count": ContextRow.trigger_count + 1,
-                    "time": ctx_stmt.excluded.time,
-                },
-            ).returning(ContextRow.id, literal_column("(xmax = 0)").label("was_insert"))
-
-            ctx_row = (await session.execute(ctx_stmt)).first()
-            assert ctx_row is not None
-            ctx_id, ctx_created = int(ctx_row.id), bool(ctx_row.was_insert)
-
-            ans_stmt = pg_insert(ContextAnswerRow).values(
-                context_id=ctx_id,
-                keywords=ans_kw_s,
-                keywords_hash=keywords_hash(ans_kw_s),
-                group_id=group_id,
-                count=1,
-                time=answer_time,
-            )
-            ans_stmt = ans_stmt.on_conflict_do_update(
-                constraint="uq_context_answer_ctx_group_kw",
-                set_={
-                    "count": ContextAnswerRow.count + 1,
-                    "time": ans_stmt.excluded.time,
-                },
-            ).returning(ContextAnswerRow.id, literal_column("(xmax = 0)").label("was_insert"))
-
-            ans_row = (await session.execute(ans_stmt)).first()
-            assert ans_row is not None
-            ans_id, answer_created = int(ans_row.id), bool(ans_row.was_insert)
-
-            if answer_created or append_on_existing:
-                await session.execute(insert(ContextAnswerMessageRow).values(answer_id=ans_id, message=msg_s))
-
-            await session.commit()
-            if ctx_created:
-                await clear_reply_query_snapshot_cache(keywords)
-            return ctx_created
-
-    async def replace_answers(self, keywords: str, answers: list[Answer], clear_time: int) -> None:
-        khash = keywords_hash(keywords)
-        async with get_session() as session:
-            ctx_result = await session.execute(select(ContextRow).where(ContextRow.keywords_hash == khash))
-            ctx_row = ctx_result.scalar_one_or_none()
-            if ctx_row is None:
-                return
-
-            ctx_id = int(ctx_row.id)
-            await session.execute(text("SELECT pg_advisory_xact_lock(:key)"), {"key": ctx_id})
-
-            kept_ids: list[int] = []
-            for answer in answers:
-                kw = _s(answer.keywords) or ""
-                ans_stmt = (
-                    pg_insert(ContextAnswerRow)
-                    .values(
-                        context_id=ctx_id,
-                        keywords=kw,
-                        keywords_hash=keywords_hash(kw),
-                        group_id=answer.group_id,
-                        count=answer.count,
-                        time=answer.time,
-                    )
-                    .on_conflict_do_update(
-                        constraint="uq_context_answer_ctx_group_kw",
-                        set_={
-                            "keywords": kw,
-                            "count": answer.count,
-                            "time": answer.time,
-                        },
-                    )
-                    .returning(ContextAnswerRow.id)
-                )
-                ans_id = int((await session.execute(ans_stmt)).scalar_one())
-                kept_ids.append(ans_id)
-
-                await session.execute(
-                    delete(ContextAnswerMessageRow).where(ContextAnswerMessageRow.answer_id == ans_id)
-                )
-                msg_rows = [
-                    ContextAnswerMessageRow(answer_id=ans_id, message=_s(message) or "") for message in answer.messages
-                ]
-                for offset in range(0, len(msg_rows), _MSG_BATCH):
-                    session.add_all(msg_rows[offset : offset + _MSG_BATCH])
-                    await session.flush()
-
-            await delete_context_answer_orphans(session, ctx_id=ctx_id, kept_ids=kept_ids)
-
-            ctx_row.clear_time = clear_time
-            await session.commit()
-        await clear_reply_query_snapshot_cache(keywords)
-
-    async def append_ban(self, keywords: str, ban: Ban) -> None:
-        khash = keywords_hash(keywords)
-        async with get_session() as session:
-            ctx_result = await session.execute(select(ContextRow.id).where(ContextRow.keywords_hash == khash))
-            ctx_id = ctx_result.scalar_one_or_none()
-            if ctx_id is None:
-                return
-
-            await session.execute(
-                insert(ContextBanRow).values(
-                    context_id=ctx_id,
-                    keywords=_s(ban.keywords) or "",
-                    group_id=ban.group_id,
-                    reason=_s(ban.reason) or "",
-                    time=ban.time,
-                )
-            )
-            await session.commit()
-        await clear_reply_query_snapshot_cache(keywords)
-
-    async def find_ban_reply_target(self, group_id: int, reply_message: str) -> tuple[str, str] | None:
-        async with get_session(read_only=True) as session:
-            result = await session.execute(
-                select(ContextRow.keywords, ContextAnswerRow.keywords)
-                .join(ContextAnswerRow, ContextAnswerRow.context_id == ContextRow.id)
-                .join(ContextAnswerMessageRow, ContextAnswerMessageRow.answer_id == ContextAnswerRow.id)
-                .where(
-                    ContextAnswerRow.group_id == int(group_id),
-                    ContextAnswerMessageRow.message == _s(reply_message),
-                )
-                .order_by(ContextAnswerRow.time.desc(), ContextAnswerMessageRow.id.desc())
-                .limit(1)
-            )
-            row = result.one_or_none()
-            if row is None:
-                return None
-            pre_keywords, reply_keywords = row
-            return str(pre_keywords), str(reply_keywords)
-
-    async def list_answers_for_group_since(self, group_id: int, cutoff_time: int) -> list[Answer]:
-        from pallas.core.foundation.db.modules import Answer
-
-        async with get_session(read_only=True) as session:
-            result = await session.execute(
-                select(ContextAnswerRow)
-                .where(
-                    ContextAnswerRow.group_id == int(group_id),
-                    ContextAnswerRow.time >= int(cutoff_time),
-                )
-                .options(selectinload(ContextAnswerRow.messages))
-            )
-            rows = list(result.scalars().all())
-        return [
-            Answer(
-                keywords=str(row.keywords),
-                group_id=int(row.group_id),
-                count=int(row.count),
-                time=int(row.time),
-                messages=[str(msg.message) for msg in row.messages],
-            )
-            for row in rows
-        ]
-
-
-def row_to_message(row: MessageRow) -> Message:
-    from pallas.core.foundation.db.modules import Message
-
-    return Message.model_construct(
-        group_id=int(row.group_id),
-        user_id=int(row.user_id),
-        bot_id=int(row.bot_id),
-        raw_message=str(row.raw_message),
-        is_plain_text=bool(row.is_plain_text),
-        plain_text=str(row.plain_text),
-        keywords=str(row.keywords),
-        sender_name=str(row.sender_name or ""),
-        message_id=int(row.message_id) if row.message_id is not None else None,
-        reply_to_message_id=int(row.reply_to_message_id) if row.reply_to_message_id is not None else None,
-        time=int(row.time),
-    )
-
-
-class PgMessageRepository:
-    # MessageRow 有 8 列，asyncpg 单语句参数上限 32767，保守取 4000 行/批
-    _BULK_BATCH_SIZE = 4000
-
-    async def find_recent_in_group(
-        self,
-        group_id: int,
-        *,
-        before_time: int | None = None,
-        before_message_id: int | None = None,
-        user_id: int | None = None,
-        limit: int = 8,
-    ) -> list[Message]:
-        cap = max(1, min(int(limit), 32))
-        stmt = select(MessageRow).where(MessageRow.group_id == int(group_id))
-        if before_time is not None:
-            if before_message_id is not None:
-                stmt = stmt.where(
-                    or_(
-                        MessageRow.time < int(before_time),
-                        and_(MessageRow.time == int(before_time), MessageRow.message_id < int(before_message_id)),
-                    )
-                )
-            else:
-                stmt = stmt.where(MessageRow.time < int(before_time))
-        if user_id is not None:
-            stmt = stmt.where(MessageRow.user_id == int(user_id))
-        stmt = stmt.order_by(MessageRow.time.desc(), MessageRow.message_id.desc()).limit(cap)
-        async with get_session(read_only=True) as session:
-            result = await session.execute(stmt)
-            rows = list(result.scalars().all())
-        rows.reverse()
-        return [row_to_message(r) for r in rows]
-
-    async def list_group_messages_after(
-        self,
-        group_id: int,
-        *,
-        after_time: int,
-        after_message_id: int | None = None,
-        limit: int = 2000,
-    ) -> list[Message]:
-        cap = max(1, min(int(limit), 4096))
-        stmt = select(MessageRow).where(MessageRow.group_id == int(group_id))
-        if after_message_id is None:
-            stmt = stmt.where(MessageRow.time > int(after_time))
-        else:
-            stmt = stmt.where(
-                or_(
-                    MessageRow.time > int(after_time),
-                    and_(MessageRow.time == int(after_time), MessageRow.message_id > int(after_message_id)),
-                )
-            )
-        stmt = stmt.order_by(MessageRow.time.asc(), MessageRow.message_id.asc()).limit(cap)
-        async with get_session(read_only=True) as session:
-            result = await session.execute(stmt)
-            rows = list(result.scalars().all())
-        return [row_to_message(r) for r in rows]
-
-    async def find_recent_distinct_in_group(
-        self,
-        group_id: int,
-        *,
-        before_time: int | None = None,
-        before_message_id: int | None = None,
-        since_time: int | None = None,
-        user_id: int | None = None,
-        limit: int = 128,
-    ) -> list[Message]:
-        cap = max(1, min(int(limit), 256))
-        partition_key = case((MessageRow.message_id.is_(None), MessageRow.id), else_=MessageRow.message_id)
-        ranked = select(
-            MessageRow,
-            func
-            .row_number()
-            .over(partition_by=partition_key, order_by=(MessageRow.time.desc(), MessageRow.id.desc()))
-            .label("_message_rank"),
-        ).where(MessageRow.group_id == int(group_id))
-        if before_time is not None:
-            if before_message_id is not None:
-                ranked = ranked.where(
-                    or_(
-                        MessageRow.time < int(before_time),
-                        and_(MessageRow.time == int(before_time), MessageRow.message_id < int(before_message_id)),
-                    )
-                )
-            else:
-                ranked = ranked.where(MessageRow.time < int(before_time))
-        if since_time is not None:
-            ranked = ranked.where(MessageRow.time >= int(since_time))
-        if user_id is not None:
-            ranked = ranked.where(MessageRow.user_id == int(user_id))
-        ranked_subquery = ranked.subquery()
-        ranked_message = aliased(MessageRow, ranked_subquery)
-        stmt = (
-            select(ranked_message)
-            .where(ranked_subquery.c._message_rank == 1)
-            .order_by(
-                ranked_message.time.desc(),
-                ranked_message.message_id.desc().nullslast(),
-                ranked_message.id.desc(),
-            )
-            .limit(cap)
-        )
-        async with get_session(read_only=True) as session:
-            result = await session.execute(stmt)
-            rows = list(result.scalars().all())
-        rows.reverse()
-        return [row_to_message(r) for r in rows]
-
-    async def find_by_message_ids(self, group_id: int, message_ids: list[int]) -> list[Message]:
-        # 注意 QQ 新版 message_id 可能是负数，isdigit() 不认负号会误过滤，导致引用图查不到。
-        ids = {int(item) for item in message_ids if str(item or "").strip().lstrip("-").isdigit() and item is not None}
-        if not ids:
-            return []
-        stmt = (
-            select(MessageRow).where(MessageRow.group_id == int(group_id)).where(MessageRow.message_id.in_(list(ids)))
-        )
-        async with get_session(read_only=True) as session:
-            result = await session.execute(stmt)
-            rows = list(result.scalars().all())
-        return [row_to_message(r) for r in rows]
-
-    async def list_recent_group_ids_for_bot(
-        self,
-        bot_id: int,
-        *,
-        since_time: int,
-        limit: int = 128,
-    ) -> list[int]:
-        cap = max(1, min(int(limit), 512))
-        stmt = (
-            select(MessageRow.group_id)
-            .where(MessageRow.bot_id == int(bot_id))
-            .where(MessageRow.time >= int(since_time))
-            .distinct()
-            .order_by(MessageRow.group_id)
-            .limit(cap)
-        )
-        async with get_session(read_only=True) as session:
-            result = await session.execute(stmt)
-            return [int(row[0]) for row in result.all()]
-
-    async def list_recent_bot_ids_for_group(
-        self,
-        group_id: int,
-        *,
-        since_time: int,
-        limit: int = 32,
-    ) -> list[int]:
-        cap = max(1, min(int(limit), 128))
-        stmt = (
-            select(MessageRow.bot_id)
-            .where(MessageRow.group_id == int(group_id))
-            .where(MessageRow.time >= int(since_time))
-            .distinct()
-            .order_by(MessageRow.bot_id)
-            .limit(cap)
-        )
-        async with get_session(read_only=True) as session:
-            result = await session.execute(stmt)
-            return [int(row[0]) for row in result.all()]
-
-    async def bulk_insert(self, messages: list[Message]) -> None:
-        if not messages:
-            return
-        async with get_session() as session:
-            for i in range(0, len(messages), self._BULK_BATCH_SIZE):
-                batch = messages[i : i + self._BULK_BATCH_SIZE]
-                values = [
-                    {
-                        "group_id": m.group_id,
-                        "user_id": m.user_id,
-                        "bot_id": m.bot_id,
-                        "raw_message": _s(m.raw_message) or "",
-                        "is_plain_text": m.is_plain_text,
-                        "plain_text": _s(m.plain_text) or "",
-                        "keywords": _s(m.keywords) or "",
-                        "sender_name": _s(m.sender_name) or "",
-                        "message_id": m.message_id,
-                        "reply_to_message_id": m.reply_to_message_id,
-                        "suppressed_by_rage": bool(getattr(m, "suppressed_by_rage", False)),
-                        "time": m.time,
-                    }
-                    for m in batch
-                ]
-                # 幂等落库：同 (group_id, bot_id, message_id) 已存在则跳过。
-                # message_id 为 NULL 的行不参与冲突判定（PG 中 NULL <> NULL）。
-                stmt = pg_insert(MessageRow).on_conflict_do_nothing(index_elements=["group_id", "bot_id", "message_id"])
-                # 走 Core executemany，避免 ORM 构造开销
-                await session.execute(stmt, values)
-            await session.commit()
-
-
-class PgBlackListRepository:
-    async def find_all(self):
-        async with get_session() as session:
-            result = await session.execute(select(BlackListRow))
-            rows = result.scalars().all()
-            return [row_to_blacklist(r) for r in rows]
-
-    async def upsert_answers(self, group_id: int, answers: list[str]) -> None:
-        """原子 upsert，基于 group_id 唯一约束。"""
-        cleaned = _strip_null_deep(answers)
-        async with get_session() as session:
-            stmt = pg_insert(BlackListRow).values(group_id=group_id, answers=cleaned, answers_reserve=[])
-            stmt = stmt.on_conflict_do_update(
-                index_elements=["group_id"],
-                set_={"answers": stmt.excluded.answers},
-            )
-            await session.execute(stmt)
-            await session.commit()
-
-    async def upsert_answers_reserve(self, group_id: int, answers: list[str]) -> None:
-        cleaned = _strip_null_deep(answers)
-        async with get_session() as session:
-            stmt = pg_insert(BlackListRow).values(group_id=group_id, answers=[], answers_reserve=cleaned)
-            stmt = stmt.on_conflict_do_update(
-                index_elements=["group_id"],
-                set_={"answers_reserve": stmt.excluded.answers_reserve},
-            )
-            await session.execute(stmt)
-            await session.commit()
-
-    async def upsert_many_blacklist(self, entries: list[BlackList]) -> None:
-        """单事务批量 upsert 多群黑名单，避免 shutdown 收尾时逐群串行 commit。"""
-        if not entries:
-            return
-        stmt = pg_insert(BlackListRow).values([
-            {
-                "group_id": e.group_id,
-                "answers": _strip_null_deep(list(e.answers)),
-                "answers_reserve": _strip_null_deep(list(e.answers_reserve)),
-            }
-            for e in entries
-        ])
-        stmt = stmt.on_conflict_do_update(
-            index_elements=["group_id"],
-            set_={"answers": stmt.excluded.answers, "answers_reserve": stmt.excluded.answers_reserve},
-        )
-        async with get_session() as session:
-            await session.execute(stmt)
-            await session.commit()
-
-
-_CONFIG_TABLE_MAP: dict[str, tuple[type, str]] = {
-    "bot_config": (BotConfigRow, "account"),
-    "group_config": (GroupConfigRow, "group_id"),
-    "user_config": (UserConfigRow, "user_id"),
-}
-
-
-def row_to_admin_member(row: AdminMemberRow):
-    from pallas.core.foundation.db.modules import AdminMember
-
-    return AdminMember.model_construct(
-        id=row.id,
-        scope=row.scope,
-        bot_id=row.bot_id,
-        user_id=row.user_id,
-        note=row.note,
-        created_at=row.created_at,
-        updated_at=row.updated_at,
-    )
-
-
-def row_to_pallas_acl(row: PallasACLRow):
-    from pallas.core.foundation.db.modules import PallasACL
-
-    return PallasACL.model_construct(
-        id=row.id,
-        role=row.role,
-        subject=row.subject,
-        action=row.action,
-        target_scope=row.target_scope,
-        target=row.target,
-        effect=row.effect,
-        priority=row.priority,
-        source=row.source,
-        created_at=row.created_at,
-        updated_at=row.updated_at,
-    )
-
-
-class PgAdminRepository:
-    """PG 版 AdminRepository。"""
-
-    async def is_admin(self, user_id: int, *, bot_id: int | None = None) -> bool:
-        async with get_session(read_only=True) as session:
-            stmt = select(AdminMemberRow.id).where(AdminMemberRow.user_id == int(user_id))
-            if bot_id is not None:
-                stmt = stmt.where(
-                    or_(
-                        AdminMemberRow.scope == "bot",
-                        AdminMemberRow.scope == "all",
-                    )
-                ).where(
-                    or_(
-                        AdminMemberRow.scope == "all",
-                        AdminMemberRow.bot_id == int(bot_id),
-                    )
-                )
-            else:
-                stmt = stmt.where(AdminMemberRow.scope == "all")
-            row = (await session.execute(stmt.limit(1))).scalar_one_or_none()
-            return row is not None
-
-    async def upsert_member(
-        self,
-        *,
-        user_id: int,
-        scope: str,
-        bot_id: int | None = None,
-        note: str | None = None,
-    ) -> Any:
-        now = int(time.time())
-        scope_norm = "bot" if scope not in ("bot", "all") else scope
-        bot_id_norm = int(bot_id) if scope_norm == "bot" and bot_id is not None else None
-        async with get_session() as session:
-            stmt = pg_insert(AdminMemberRow).values(
-                scope=scope_norm,
-                bot_id=bot_id_norm,
-                user_id=int(user_id),
-                note=note,
-                created_at=now,
-                updated_at=now,
-            )
-            stmt = stmt.on_conflict_do_update(
-                index_elements=["scope", "bot_id", "user_id"],
-                set_={
-                    "updated_at": stmt.excluded.updated_at,
-                    "note": stmt.excluded.note,
-                },
-            )
-            row = (await session.execute(stmt.returning(AdminMemberRow))).scalar_one()
-            await session.commit()
-            return row_to_admin_member(row)
-
-    async def remove_member(
-        self,
-        *,
-        user_id: int,
-        scope: str,
-        bot_id: int | None = None,
-    ) -> int:
-        scope_norm = "bot" if scope not in ("bot", "all") else scope
-        bot_id_norm = int(bot_id) if scope_norm == "bot" and bot_id is not None else None
-        async with get_session() as session:
-            result = await session.execute(
-                delete(AdminMemberRow).where(
-                    AdminMemberRow.scope == scope_norm,
-                    AdminMemberRow.bot_id == bot_id_norm,
-                    AdminMemberRow.user_id == int(user_id),
-                )
-            )
-            await session.commit()
-            return int(result.rowcount or 0)
-
-    async def delete_member(self, member_id: Any) -> int:
-        try:
-            rid = int(member_id)
-        except (TypeError, ValueError):
-            return 0
-        async with get_session() as session:
-            result = await session.execute(delete(AdminMemberRow).where(AdminMemberRow.id == rid))
-            await session.commit()
-            return int(result.rowcount or 0)
-
-    async def list_members(
-        self,
-        *,
-        scope: str | None = None,
-        bot_id: int | None = None,
-    ) -> list[Any]:
-        async with get_session(read_only=True) as session:
-            stmt = select(AdminMemberRow)
-            if scope is not None:
-                stmt = stmt.where(AdminMemberRow.scope == scope)
-            if bot_id is not None:
-                stmt = stmt.where(AdminMemberRow.bot_id == int(bot_id))
-            rows = (await session.execute(stmt)).scalars().all()
-            return [row_to_admin_member(r) for r in rows]
-
-    async def has_user(self, user_id: int) -> bool:
-        async with get_session(read_only=True) as session:
-            stmt = select(AdminMemberRow.id).where(AdminMemberRow.user_id == int(user_id)).limit(1)
-            row = (await session.execute(stmt)).scalar_one_or_none()
-            return row is not None
-
-    async def list_admin_user_ids(self, *, bot_id: int | None) -> list[int]:
-        async with get_session(read_only=True) as session:
-            if bot_id is None:
-                stmt = select(AdminMemberRow.user_id).where(AdminMemberRow.scope == "all")
-            else:
-                stmt = select(AdminMemberRow.user_id).where(
-                    or_(
-                        AdminMemberRow.scope == "all",
-                        and_(
-                            AdminMemberRow.scope == "bot",
-                            AdminMemberRow.bot_id == int(bot_id),
-                        ),
-                    )
-                )
-            rows = (await session.execute(stmt)).scalars().all()
-            return [int(r) for r in rows if r is not None]
-
-
-class PgAclRepository:
-    """PG 版 AclRepository。"""
-
-    async def list_rules(
-        self,
-        *,
-        action: str | None = None,
-        target: str | None = None,
-        role: str | None = None,
-        subject: str | None = None,
-    ) -> list[Any]:
-        async with get_session(read_only=True) as session:
-            stmt = select(PallasACLRow)
-            if action is not None:
-                stmt = stmt.where(PallasACLRow.action == action)
-            if target is not None:
-                stmt = stmt.where(PallasACLRow.target == target)
-            if role is not None:
-                stmt = stmt.where(PallasACLRow.role == role)
-            if subject is not None:
-                stmt = stmt.where(PallasACLRow.subject == subject)
-            rows = (await session.execute(stmt)).scalars().all()
-            return [row_to_pallas_acl(r) for r in rows]
-
-    async def list_all(self) -> list[Any]:
-        async with get_session(read_only=True) as session:
-            rows = (await session.execute(select(PallasACLRow))).scalars().all()
-            return [row_to_pallas_acl(r) for r in rows]
-
-    async def list_matching_rules(
-        self,
-        *,
-        action: str,
-        target: str | None = None,
-    ) -> list[Any]:
-        async with get_session(read_only=True) as session:
-            stmt = select(PallasACLRow).where(PallasACLRow.action == action)
-            if target is not None:
-                stmt = stmt.where(
-                    or_(
-                        PallasACLRow.target == "*",
-                        PallasACLRow.target == target,
-                    )
-                )
-            rows = (await session.execute(stmt)).scalars().all()
-            return [row_to_pallas_acl(r) for r in rows]
-
-    async def upsert_rule(
-        self,
-        *,
-        role: str,
-        subject: str | None,
-        action: str,
-        target_scope: str,
-        target: str,
-        effect: str,
-        priority: int,
-        source: str,
-    ) -> Any:
-        now = int(time.time())
-        async with get_session() as session:
-            stmt = pg_insert(PallasACLRow).values(
-                role=role,
-                subject=subject,
-                action=action,
-                target_scope=target_scope,
-                target=target,
-                effect=effect,
-                priority=int(priority),
-                source=source,
-                created_at=now,
-                updated_at=now,
-            )
-            stmt = stmt.on_conflict_do_update(
-                index_elements=[
-                    "role",
-                    "subject",
-                    "action",
-                    "target_scope",
-                    "target",
-                ],
-                set_={
-                    "effect": stmt.excluded.effect,
-                    "priority": stmt.excluded.priority,
-                    "source": stmt.excluded.source,
-                    "updated_at": stmt.excluded.updated_at,
-                },
-            )
-            row = (await session.execute(stmt.returning(PallasACLRow))).scalar_one()
-            await session.commit()
-            return row_to_pallas_acl(row)
-
-    async def delete_rule(self, rule_id: Any) -> int:
-        try:
-            rid = int(rule_id)
-        except (TypeError, ValueError):
-            return 0
-        async with get_session() as session:
-            result = await session.execute(delete(PallasACLRow).where(PallasACLRow.id == rid))
-            await session.commit()
-            return int(result.rowcount or 0)
-
-    async def delete_by_signature(
-        self,
-        *,
-        role: str,
-        subject: str | None,
-        action: str,
-        target_scope: str,
-        target: str,
-    ) -> int:
-        async with get_session() as session:
-            result = await session.execute(
-                delete(PallasACLRow).where(
-                    PallasACLRow.role == role,
-                    PallasACLRow.subject == subject,
-                    PallasACLRow.action == action,
-                    PallasACLRow.target_scope == target_scope,
-                    PallasACLRow.target == target,
-                )
-            )
-            await session.commit()
-            return int(result.rowcount or 0)
-
-    async def list_group_block_targets(self) -> set[str]:
-        async with get_session(read_only=True) as session:
-            stmt = select(PallasACLRow.target).where(PallasACLRow.target.like("group:%"))
-            rows = (await session.execute(stmt)).scalars().all()
-            return {row for row in rows if row}
-
-    async def has_run_step(self, step: str) -> bool:
-        async with get_session(read_only=True) as session:
-            row = (
-                await session.execute(select(SchemaMigrationRow.id).where(SchemaMigrationRow.step == step).limit(1))
-            ).scalar_one_or_none()
-            return row is not None
-
-    async def mark_run_step(self, step: str) -> None:
-        now = int(time.time())
-        async with get_session() as session:
-            stmt = pg_insert(SchemaMigrationRow).values(step=step, applied_at=now)
-            stmt = stmt.on_conflict_do_update(
-                index_elements=["step"],
-                set_={"applied_at": stmt.excluded.applied_at},
-            )
-            await session.execute(stmt)
-            await session.commit()
-
-
-# ---------------------------------------------------------------------------
-# 配置 Repository 的 TTL 缓存
-# ---------------------------------------------------------------------------
-
-
-def _cfg_env(key: str, default: str) -> str:
-    try:
-        import nonebot
-
-        val = getattr(nonebot.get_driver().config, key.lower(), None)
-        if val is not None:
-            return str(val)
-    except Exception:
-        pass
-    return os.getenv(key, default)
-
-
-class _ConfigCache:
-    """
-    简单的容量 + TTL 缓存，对齐 Mongo Beanie 的 model-level cache 语义。
-    对每个 (row_class) 一个实例；key 是主键值，value 是 (row, expire_ts)
-    None 也会被缓存。
-    """
-
-    def __init__(self, ttl: float, capacity: int) -> None:
-        self._ttl = ttl
-        self._capacity = capacity
-        self._store: OrderedDict[Any, tuple[Any, float]] = OrderedDict()
-        self._lock = asyncio.Lock()
-
-    async def get(self, key: Any) -> tuple[bool, Any]:
-        """TTL 缓存查询，返回 hit 与 value。"""
-        if self._ttl <= 0 or self._capacity <= 0:
-            return False, None
-        async with self._lock:
-            item = self._store.get(key)
-            if item is None:
-                return False, None
-            value, expire_ts = item
-            if expire_ts <= time.time():
-                self._store.pop(key, None)
-                return False, None
-            self._store.move_to_end(key)
-            return True, value
-
-    async def put(self, key: Any, value: Any) -> None:
-        if self._ttl <= 0 or self._capacity <= 0:
-            return
-        async with self._lock:
-            self._store[key] = (value, time.time() + self._ttl)
-            self._store.move_to_end(key)
-            while len(self._store) > self._capacity:
-                self._store.popitem(last=False)
-
-    async def invalidate(self, key: Any) -> None:
-        async with self._lock:
-            self._store.pop(key, None)
-
-    async def clear(self) -> None:
-        async with self._lock:
-            self._store.clear()
-
-
-_CONFIG_CACHES: dict[type, _ConfigCache] = {}
-
-
-def _get_config_cache(row_class: type) -> _ConfigCache:
-    cache = _CONFIG_CACHES.get(row_class)
-    if cache is None:
-        ttl = float(_cfg_env("PG_CONFIG_CACHE_TTL", "60"))
-        capacity = int(_cfg_env("PG_CONFIG_CACHE_SIZE", "10000"))
-        cache = _ConfigCache(ttl=ttl, capacity=capacity)
-        _CONFIG_CACHES[row_class] = cache
-    return cache
-
-
-class PgConfigRepository:
-    def __init__(self, table: str, primary_key: str) -> None:
-        if table not in _CONFIG_TABLE_MAP:
-            raise ValueError(f"Unknown config table: {table}")
-        row_class, pk_field = _CONFIG_TABLE_MAP[table]
-        # primary_key 由工厂函数传入，
-        # 这里做一致性断言，避免静默与 _CONFIG_TABLE_MAP 失同步。
-        if primary_key != pk_field:
-            raise ValueError(f"primary_key {primary_key!r} 与 {table} 登记的主键 {pk_field!r} 不一致")
-        self._row_class, self._pk_field = row_class, pk_field
-        self._cache = _get_config_cache(self._row_class)
-
-    async def get(self, key_id: int, *, ignore_cache: bool = False) -> Any | None:
-        if not ignore_cache:
-            hit, value = await self._cache.get(key_id)
-            if hit:
-                return value
-        async with get_session(read_only=True) as session:
-            result = await session.execute(
-                select(self._row_class).where(getattr(self._row_class, self._pk_field) == key_id)
-            )
-            row = result.scalar_one_or_none()
-            if row is not None:
-                session.expunge(row)
-        await self._cache.put(key_id, row)
-        return row
-
-    async def list_all(self) -> list[Any]:
-        async with get_session(read_only=True) as session:
-            rows = list((await session.execute(select(self._row_class))).scalars().all())
-            for row in rows:
-                session.expunge(row)
-        return rows
-
-    async def get_or_create(self, key_id: int, **defaults: Any) -> tuple[Any, bool]:
-        async with get_session() as session:
-            result = await session.execute(
-                select(self._row_class).where(getattr(self._row_class, self._pk_field) == key_id)
-            )
-            row = result.scalar_one_or_none()
-            if row is not None:
-                session.expunge(row)
-                await self._cache.put(key_id, row)
-                return row, False
-            try:
-                new_row = self._row_class(**{self._pk_field: key_id, **_strip_null_deep(defaults)})
-                session.add(new_row)
-                await session.commit()
-            except IntegrityError:
-                # 并发下已被其他 writer 插入，回源拿最新行
-                await session.rollback()
-                result = await session.execute(
-                    select(self._row_class).where(getattr(self._row_class, self._pk_field) == key_id)
-                )
-                existing = result.scalar_one_or_none()
-                if existing is not None:
-                    session.expunge(existing)
-                await self._cache.put(key_id, existing)
-                return existing, False
-            session.expunge(new_row)
-            await self._cache.put(key_id, new_row)
-            return new_row, True
-
-    async def upsert_field(self, key_id: int, field: str, value: Any) -> None:
-        """字段级 upsert，基于主键 ON CONFLICT 原子化。"""
-        cleaned_value = _strip_null_deep(value)
-        async with get_session() as session:
-            stmt = pg_insert(self._row_class).values(**{self._pk_field: key_id, field: cleaned_value})
-            stmt = stmt.on_conflict_do_update(
-                index_elements=[self._pk_field],
-                set_={field: getattr(stmt.excluded, field)},
-            )
-            await session.execute(stmt)
-            await session.commit()
-        await self._cache.invalidate(key_id)
-
-    async def upsert_fields(self, key_id: int, fields: dict[str, Any]) -> None:
-        """批量字段级 upsert"""
-        if not fields:
-            return
-        cleaned = {k: _strip_null_deep(v) for k, v in fields.items()}
-        async with get_session() as session:
-            stmt = pg_insert(self._row_class).values(**{self._pk_field: key_id, **cleaned})
-            stmt = stmt.on_conflict_do_update(
-                index_elements=[self._pk_field],
-                set_={k: getattr(stmt.excluded, k) for k in cleaned},
-            )
-            await session.execute(stmt)
-            await session.commit()
-        await self._cache.invalidate(key_id)
-
-    async def invalidate_cache(self) -> None:
-        await self._cache.clear()
-
-
-def image_cache_has_blob_clause():
-    """迁移期兼容：文件化后走 blob_path，存量旧行 blob_data 仍有效。"""
-    return or_(
-        ImageCacheRow.blob_path.is_not(None),
-        and_(ImageCacheRow.blob_data.is_not(None), ImageCacheRow.blob_data != b""),
-    )
-
-
-def image_cache_persist_blob(cache: ImageCache) -> dict[str, Any]:
-    """blob_data 有值时写文件，返回入库用的 content_hash/blob_path/blob_size。"""
-    data = cache.blob_data
-    if not data:
-        return {}
-    content_hash = cache.content_hash or hashlib.sha256(data).hexdigest()
-    return {
-        "content_hash": content_hash,
-        "blob_path": str(image_blob_rel_path(content_hash)),
-        "blob_size": write_image_blob(content_hash, data),
-    }
-
-
-async def image_cache_cleanup_blobs(hashes: list[str | None]) -> None:
-    present = {h for h in hashes if h}
-    if not present:
-        return
-    async with get_session(read_only=True) as session:
-        rows = (
-            (
-                await session.execute(
-                    select(ImageCacheRow.content_hash).where(ImageCacheRow.content_hash.in_(tuple(present))).distinct()
-                )
-            )
-            .scalars()
-            .all()
-        )
-    referenced = set(rows)
-    for content_hash in present - referenced:
-        delete_image_blob(content_hash)
-
-
-async def image_cache_fill_blob(cache: ImageCache | None) -> ImageCache | None:
-    """返回前把文件 blob 装入内存，兼容直接调 repo 的消费方与迁移期旧行。"""
-    if cache is None or cache.blob_data or not cache.blob_path:
-        return cache
-    cache.blob_data = await asyncio.to_thread(read_image_blob_at, cache.blob_path)
-    return cache
-
-
-class PgImageCacheRepository:
-    async def find_by_cq_code(self, cq_code: str) -> ImageCache | None:
-        async with get_session(read_only=True) as session:
-            result = await session.execute(select(ImageCacheRow).where(ImageCacheRow.cq_code == cq_code))
-            row = result.scalar_one_or_none()
-            return await image_cache_fill_blob(row_to_image_cache(row) if row else None)
-
-    async def find_by_content_hash(self, content_hash: str) -> ImageCache | None:
-        async with get_session(read_only=True) as session:
-            stmt = (
-                select(ImageCacheRow)
-                .where(ImageCacheRow.content_hash == content_hash)
-                .order_by(ImageCacheRow.date.desc(), ImageCacheRow.id.desc())
-                .limit(1)
-            )
-            row = (await session.execute(stmt)).scalars().first()
-            return await image_cache_fill_blob(row_to_image_cache(row) if row else None)
-
-    async def find_by_url(self, url: str) -> ImageCache | None:
-        async with get_session(read_only=True) as session:
-            stmt = (
-                select(ImageCacheRow)
-                .where(ImageCacheRow.cq_code.contains(url), image_cache_has_blob_clause())
-                .order_by(ImageCacheRow.date.desc(), ImageCacheRow.id.desc())
-                .limit(1)
-            )
-            row = (await session.execute(stmt)).scalars().first()
-            return await image_cache_fill_blob(row_to_image_cache(row) if row else None)
-
-    async def bind_content_hash(self, cq_code: str, content_hash: str) -> None:
-        async with get_session() as session:
-            await session.execute(
-                update(ImageCacheRow).where(ImageCacheRow.cq_code == cq_code).values(content_hash=content_hash)
-            )
-            await session.commit()
-
-    async def find_latest_with_blob(self) -> ImageCache | None:
-        async with get_session(read_only=True) as session:
-            stmt = (
-                select(ImageCacheRow)
-                .where(image_cache_has_blob_clause())
-                .order_by(ImageCacheRow.date.desc(), ImageCacheRow.id.desc())
-                .limit(1)
-            )
-            row = (await session.execute(stmt)).scalar_one_or_none()
-            return await image_cache_fill_blob(row_to_image_cache(row) if row else None)
-
-    async def find_recent_with_blob(self, limit: int) -> list[ImageCache]:
-        async with get_session(read_only=True) as session:
-            stmt = (
-                select(ImageCacheRow)
-                .where(image_cache_has_blob_clause())
-                .order_by(ImageCacheRow.date.desc(), ImageCacheRow.id.desc())
-                .limit(max(1, int(limit)))
-            )
-            rows = (await session.execute(stmt)).scalars().all()
-        caches = [row_to_image_cache(row) for row in rows]
-        return [cache for cache in await asyncio.gather(*(image_cache_fill_blob(c) for c in caches)) if cache]
-
-    async def insert(self, cache: ImageCache) -> None:
-        """并发下相同 cq_code 的第二次 insert 等价为 no-op。"""
-        blob = image_cache_persist_blob(cache)
-        values: dict[str, Any] = {
-            "cq_code": _s(cache.cq_code) or "",
-            "content_hash": _s(blob.get("content_hash") or cache.content_hash),
-            "blob_path": blob.get("blob_path"),
-            "blob_size": blob.get("blob_size"),
-            "ref_times": cache.ref_times,
-            "date": cache.date,
-        }
-        # 迁移期：blob 尚未文件化时沿用旧 bytea 列
-        if not blob and cache.blob_data:
-            values["blob_data"] = cache.blob_data
-        async with get_session() as session:
-            stmt = pg_insert(ImageCacheRow).values(values)
-            await session.execute(stmt.on_conflict_do_nothing(index_elements=["cq_code"]))
-            await session.commit()
-
-    async def save(self, cache: ImageCache) -> None:
-        """upsert 语义：存在则更新，否则插入。"""
-        blob = image_cache_persist_blob(cache)
-        values: dict[str, Any] = {
-            "cq_code": _s(cache.cq_code) or "",
-            "content_hash": _s(blob.get("content_hash") or cache.content_hash),
-            "blob_path": blob.get("blob_path"),
-            "blob_size": blob.get("blob_size"),
-            "ref_times": cache.ref_times,
-            "date": cache.date,
-        }
-        if not blob and cache.blob_data:
-            values["blob_data"] = cache.blob_data
-        async with get_session() as session:
-            stmt = pg_insert(ImageCacheRow).values(values)
-            set_ = {
-                "ref_times": stmt.excluded.ref_times,
-                "date": stmt.excluded.date,
-                "content_hash": stmt.excluded.content_hash,
-                "blob_path": stmt.excluded.blob_path,
-                "blob_size": stmt.excluded.blob_size,
-            }
-            if not blob:
-                set_["blob_data"] = stmt.excluded.blob_data
-            stmt = stmt.on_conflict_do_update(index_elements=["cq_code"], set_=set_)
-            await session.execute(stmt)
-            await session.commit()
-
-    async def touch(self, cq_code: str, *, date: int) -> None:
-        async with get_session() as session:
-            await session.execute(
-                update(ImageCacheRow)
-                .where(ImageCacheRow.cq_code == cq_code)
-                .values(ref_times=ImageCacheRow.ref_times + 1, date=int(date))
-            )
-            await session.commit()
-
-    async def delete_old(self, before_date: int) -> None:
-        while True:
-            async with get_session(read_only=True) as session:
-                rows = (
-                    await session.execute(
-                        select(ImageCacheRow.id, ImageCacheRow.content_hash)
-                        .where(ImageCacheRow.date < int(before_date))
-                        .limit(1000)
-                    )
-                ).all()
-            if not rows:
-                break
-            ids = [int(row[0]) for row in rows]
-            async with get_session() as session:
-                await session.execute(delete(ImageCacheRow).where(ImageCacheRow.id.in_(ids)))
-                await session.commit()
-            await image_cache_cleanup_blobs([str(row[1]) for row in rows])
-            if len(rows) < 1000:
-                break
-
-    async def delete_low_ref(self, ref_threshold: int) -> None:
-        while True:
-            async with get_session(read_only=True) as session:
-                rows = (
-                    await session.execute(
-                        select(ImageCacheRow.id, ImageCacheRow.content_hash)
-                        .where(ImageCacheRow.ref_times < ref_threshold)
-                        .limit(1000)
-                    )
-                ).all()
-            if not rows:
-                break
-            ids = [int(row[0]) for row in rows]
-            async with get_session() as session:
-                await session.execute(delete(ImageCacheRow).where(ImageCacheRow.id.in_(ids)))
-                await session.commit()
-            await image_cache_cleanup_blobs([str(row[1]) for row in rows])
-            if len(rows) < 1000:
-                break
-
-    async def prune(self, policy: ImageCachePrunePolicy) -> ImageCachePruneResult:
-        from pallas.core.foundation.db.repository import ImageCachePruneResult
-
-        batch_size = max(1, int(policy.batch_size))
-        max_blob_bytes = max(0, int(policy.max_blob_bytes))
-        deleted_rows = 0
-        deleted_blob_bytes = 0
-        blob_size_expr = func.coalesce(
-            ImageCacheRow.blob_size,
-            func.pg_column_size(ImageCacheRow.blob_data),
-            0,
-        )
-
-        async def total_blob_bytes() -> int:
-            async with get_session(read_only=True) as session:
-                stmt = select(func.coalesce(func.sum(blob_size_expr), 0))
-                return int((await session.execute(stmt)).scalar_one())
-
-        async def delete_batch(where_clause, *, size_limit: int | None = None) -> tuple[int, int]:
-            async with get_session(read_only=True) as session:
-                stmt = (
-                    select(ImageCacheRow.id, ImageCacheRow.content_hash, blob_size_expr)
-                    .where(where_clause)
-                    .order_by(
-                        case((ImageCacheRow.ref_times <= 1, 0), else_=1),
-                        ImageCacheRow.date,
-                        ImageCacheRow.id,
-                    )
-                    .limit(batch_size)
-                )
-                rows = list((await session.execute(stmt)).all())
-                if size_limit is not None:
-                    selected = []
-                    selected_bytes = 0
-                    for row in rows:
-                        selected.append(row)
-                        selected_bytes += int(row[2])
-                        if selected_bytes >= size_limit:
-                            break
-                    rows = selected
-                if not rows:
-                    return 0, 0
-                ids = [int(row[0]) for row in rows]
-                hashes = [str(row[1]) for row in rows]
-            async with get_session() as session:
-                await session.execute(delete(ImageCacheRow).where(ImageCacheRow.id.in_(ids)))
-                await session.commit()
-            await image_cache_cleanup_blobs(hashes)
-            return len(rows), sum(int(row[2]) for row in rows)
-
-        expired = or_(
-            ImageCacheRow.date < int(policy.absolute_before),
-            and_(ImageCacheRow.date < int(policy.single_use_before), ImageCacheRow.ref_times <= 1),
-        )
-        while True:
-            rows, blob_bytes = await delete_batch(expired)
-            deleted_rows += rows
-            deleted_blob_bytes += blob_bytes
-            if rows < batch_size:
-                break
-
-        remaining_blob_bytes = await total_blob_bytes()
-        while remaining_blob_bytes > max_blob_bytes:
-            rows, blob_bytes = await delete_batch(
-                literal_column("true"),
-                size_limit=remaining_blob_bytes - max_blob_bytes,
-            )
-            if not rows:
-                break
-            deleted_rows += rows
-            deleted_blob_bytes += blob_bytes
-            remaining_blob_bytes = max(0, remaining_blob_bytes - blob_bytes)
-
-        return ImageCachePruneResult(
-            deleted_rows=deleted_rows,
-            deleted_blob_bytes=deleted_blob_bytes,
-            remaining_blob_bytes=remaining_blob_bytes,
-        )
-
-
-async def vacuum_message_table() -> None:
-    """语料扫库批量删除后回收 message 表死元组，缓解 autovacuum 滞后导致的索引膨胀。
-
-    VACUUM 不能运行在事务块内，需借用 AUTOCOMMIT 连接；失败仅降级不影响主流程。
-    """
-    if _engine is None:
-        return
-    try:
-        async with _engine.connect() as conn:
-            await conn.execution_options(isolation_level="AUTOCOMMIT").exec_driver_sql("VACUUM message")
-    except Exception as exc:
-        logger.warning("vacuum message failed, degraded: {}", exc)
